==> Scripts/2 Introduction and scale-up description (Fig 1).Rmd <==
---
title: "2 Introduction and scale-up description (Fig 1)"
author: "Martina Pesce & Edward P. K. Parker"
date: "2023-11-30"
output: html_document
editor_options: 
  chunk_output_type: console
---

#### Inputs
- Processed/Ineq_merged.csv: Combined country-level income data
- Input/Vaccine_data/owid-covid-data.csv: Our World in Data COVID-19 database
- Input/Vaccine_data/owid_vaccine_plat.csv: Our World in Data COVID-19 vaccine platforms
- Input/Vaccine_data/vaccination-data.csv: WHO COVID-19 vaccine database

#### Outputs
- Processed/Ineq_merged_with_vaccine_data.csv: Country-level income data with vaccine introduction and scale-up metrics 
- Processed/Ineq_merged_with_vaccine_data_plus_priority_groups.csv: data inequalities OWID/WHO vaccine data + metrics + priority groups
- Processed/Ineq_merged_with_vaccine_priority_groups.csv: priority groups metrics

- Fig 4. Vaccine coverage in high-risk populations (health care workers and older adults) and overall. 
- Supplementary Fig. 1. Percentage of countries using different vaccine products by income status.
- Supplementary Fig. 7. Map of countries by per capita Gross National Income quartile.
- Supplementary Table 1. Vaccine introduction and scale-up milestones by income status.
- Supplementary Table 2. Vaccine coverage in high-risk populations and overall according to WHO/UNICEF database.
- Supplementary Table 3.  Age thresholds for defining older adults.


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```




```{r Libraries}
library(tidyverse)
library(readr)
library(sf)
library(rnaturalearth)
# library(RColorBrewer)
# library(lubridate)
library(countrycode)
# library(patchwork)
 
```



```{r Data base F1}

#Ineq database GNI
Ineq_merged <- read_csv("Data/Processed/Ineq_merged.csv") %>% 
  select(Country.Name:q_GNI_per_c_WB)


#Coverage information

##Our world in data 
owid_covid_data <- read_csv("Data/Input/Vaccine_data/owid-covid-data_2024-01-07.csv") %>% 
select(
    iso_code:date, total_vaccinations_per_hundred: people_fully_vaccinated_per_hundred)

# Align iso3 for Kosovo
owid_covid_data$iso_code[owid_covid_data$location=="Kosovo"]
owid_covid_data$iso_code[owid_covid_data$location=="Kosovo"] = "XKX"

##WHO
WHO_covid_data <- read_csv("Data/Input/Vaccine_data/WHO_vaccination-data-2024-01-07.csv") %>%   
  select(
    ISO3, FIRST_VACCINE_DATE
  )

##WHO health care workers and older populations
WHO_hcw_old_data <- read_csv("Data/Input/Vaccine_data/WHO_UNICEF_data/who_unicef_country_timeseries_data.csv") %>% 
  select(name_short, iso, date, cov_hcw_cps, cov_old_cps, cov_tot_cps,pop, pop_hcw, pop_old) %>% 
  mutate(date = ifelse(is.na(date), NA, paste0(date, "-01")))

WHO_hcw_old_sum_data <- read_csv("Data/Input/Vaccine_data/WHO_UNICEF_data/who_unicef_country_summary_data.csv") %>% 
  select(iso, pol_age_old, pol_age_old_source)

# merge sources
WHO_hcw_old_data = WHO_hcw_old_data %>%
  left_join(WHO_hcw_old_sum_data, by = "iso") %>%
  mutate(
    perc_HCW = pop_hcw/pop,
    perc_old = pop_old/pop
  )

#vaccines platforms
owid_vaccine_plat <- read_csv("Data/Input/Vaccine_data/owid_vaccine_plat.csv") %>%
  select(location:source_website)

# Align iso3 for Kosovo
owid_vaccine_plat$iso_code[owid_vaccine_plat$location=="Kosovo"]
owid_vaccine_plat$iso_code[owid_vaccine_plat$location=="Kosovo"] = "XKX"
```


#Data transformation for descriptive analysis

```{r}
#OWID data country, milestone achievement date
coverage <- owid_covid_data %>%
  group_by(location) %>%
  mutate(
    owid_vac_intro = ifelse(row_number() == which(total_vaccinations_per_hundred > 0)[1], as.character(date), NA),
    owid_vac_F40 = ifelse(row_number() == which(people_fully_vaccinated_per_hundred > 39.99)[1], as.character(date), NA),
    owid_vac_F70 = ifelse(row_number() == which(people_fully_vaccinated_per_hundred > 69.99)[1], as.character(date), NA)
  ) %>%
  filter(!is.na(owid_vac_intro) | !is.na(owid_vac_F40) | !is.na(owid_vac_F70)) %>% 
  select(!(date:people_fully_vaccinated_per_hundred)) %>% 
  fill(owid_vac_intro:owid_vac_F70)%>%
  slice_tail() %>%  
  ungroup()

# Merge data inequalities/OWID vac data / WHO vaccine data and calculate minimum vaccine campaign date (OWID vs WHO)
Ineq_vaccine_coverage <- Ineq_merged %>%
  left_join(coverage, by = c("Country.Code" = "iso_code")) %>%
  left_join(WHO_covid_data, by = c("Country.Code" = "ISO3")) %>%
  mutate(
    who_vac_intro = as.Date(as.character(FIRST_VACCINE_DATE)),
    owid_vac_intro = as.Date(owid_vac_intro)
  ) %>%
  mutate(
    min_vac_intro = pmin(owid_vac_intro, who_vac_intro, na.rm = TRUE),
    diff_owid_who_intro = as.numeric(owid_vac_intro - who_vac_intro)
  )%>% 
    mutate(
    diff_intro_F40 = as.numeric(as.Date(owid_vac_F40) - as.Date(min_vac_intro)),
    diff_intro_F70 = as.numeric(as.Date(owid_vac_F70) - as.Date(min_vac_intro)),
    owid_who= case_when(
      diff_owid_who_intro < 0 | (is.na(who_vac_intro) & !is.na(owid_vac_intro))~ "owid",
      diff_owid_who_intro > 0 | (is.na(owid_vac_intro) & !is.na(who_vac_intro)) ~ "who",
      diff_owid_who_intro == 0 ~ "equal",
      TRUE ~ "no date"
    )
    ) %>%
  select(-FIRST_VACCINE_DATE, -location)

#Number countries with date in OWID or WHO
table(Ineq_vaccine_coverage$owid_who)

# Countries with date in both 
table(!is.na(
  Ineq_vaccine_coverage[Ineq_vaccine_coverage$owid_who != "no date", "diff_owid_who_intro"]
))

# WHO/OWID date difference
Ineq_vaccine_coverage %>%
  summarise(
    Q1     = quantile(diff_owid_who_intro, 0.25, na.rm = TRUE),
    median = median(diff_owid_who_intro, na.rm = TRUE),
    Q3     = quantile(diff_owid_who_intro, 0.75, na.rm = TRUE),
    less_7_days = sum(Ineq_vaccine_coverage$diff_owid_who_intro < 8, na.rm = TRUE)
  )

#check owid-who
summary_countries_who_owid <- Ineq_vaccine_coverage %>%
  mutate(non_na = !is.na(diff_owid_who_intro)) %>%
  group_by(non_na) %>% 
  summarise(
    count = n(),
    median_diff = round(median(diff_owid_who_intro, na.rm = TRUE), 0),
    q25_diff = round(quantile(diff_owid_who_intro, 0.25, na.rm = TRUE), 0),
    q75_diff = round(quantile(diff_owid_who_intro, 0.75, na.rm = TRUE), 0),
    less_7_days = sum(Ineq_vaccine_coverage$diff_owid_who_intro < 8, na.rm = TRUE)
    )
```



```{r}
# + Vaccine Introduction and scale-up metrics data (for figures)

#Create groups 
Ineq_vaccine_coverage <- Ineq_vaccine_coverage %>% 
  mutate(
            V1_date_group=  case_when(
      min_vac_intro < dmy("01-01-2021")  ~ "before jan/2021",
      min_vac_intro >= dmy("01-01-2021") & min_vac_intro < dmy("01-02-2021") ~ "jan/2021",
      min_vac_intro >= dmy("01-02-2021") & min_vac_intro < dmy("01-03-2021") ~ "feb/2021",
      min_vac_intro >= dmy("01-03-2021") ~ "march/2021 +",
      TRUE ~ "not reached"
  ),
            V40_date_group=  case_when(
      owid_vac_F40 < dmy("01-06-2021")  ~ "before Jun/2021",
      owid_vac_F40 >= dmy("01-06-2021") & owid_vac_F40 < dmy("01-09-2021") ~ "Jun-Aug/2021",
      owid_vac_F40 >= dmy("01-09-2021") & owid_vac_F40 < dmy("01-12-2021") ~ "Sep-Nov/2021",
      owid_vac_F40 >= dmy("01-12-2021") ~ "Dec/2021 +",
      TRUE ~ "not reached"
  ),
            V40_group= case_when(
      diff_intro_F40 < 180  ~ "<180",
      diff_intro_F40 > 179 & diff_intro_F40 < 270 ~ "180-269",
      diff_intro_F40 > 269 & diff_intro_F40 < 360 ~ "270-359",
      diff_intro_F40 > 359 ~ "360+",
      TRUE ~ "not reached"
  ),
    V1_dummy = case_when(
      !is.na(min_vac_intro) & min_vac_intro < as.Date("2022-01-01") ~ "yes",
      TRUE ~ "no"
    ),
    V40_dummy = case_when(
      !is.na(owid_vac_F40) & owid_vac_F40 < as.Date("2022-01-01") ~ "yes",
      TRUE ~ "no"
    ),
    V70_dummy = case_when(
      !is.na(owid_vac_F70) & owid_vac_F70 < as.Date("2022-07-01") ~ "yes",
      TRUE ~ "no"
    ),
    V1_dummy_no_date = case_when(
      !is.na(min_vac_intro) ~ "yes",
      TRUE ~ "no"
    ),
    V40_dummy_no_date = case_when(
      !is.na(owid_vac_F40) ~ "yes",
      TRUE ~ "no"
    ),
    V70_dummy_no_date = case_when(
      !is.na(owid_vac_F70) ~ "yes",
      TRUE ~ "no"
    )
  )
  
```



```{r owid vacc plat}
# + Vaccine platforms (OWID) by GNIpc

# List vaccines 
unique(unlist(strsplit(paste(owid_vaccine_plat$vaccines, collapse = ', '), ", ")))

# Create col for each vacc (1/0)
Ineq_vaccine_coverage_platform <- Ineq_vaccine_coverage %>%
  left_join(owid_vaccine_plat, by=c("Country.Code" = "iso_code")) %>%
  select(-location) %>%
  mutate(
    `CIGB CIGB-66/Abdala` = ifelse(str_detect(vaccines, "Abdala"), 1, 0),
    `Cansino Ad5-nCoV` = ifelse(str_detect(vaccines, "CanSino"), 1, 0),
    `Biological E Corbevax` = ifelse(str_detect(vaccines, "Corbevax"), 1, 0),
    `Bharat Covaxin/BBV152` = ifelse(str_detect(vaccines, "Covaxin"), 1, 0),
    `Shifa Pharmed COVIran` = ifelse(str_detect(vaccines, "COVIran Barekat"), 1, 0),
    `Chumakov Center CoviVac` = ifelse(str_detect(vaccines, "KoviVac/Chumakov"), 1, 0),
    `Vector Institute EpiVacCorona` = ifelse(str_detect(vaccines, "EpiVacCorona"), 1, 0),
    `Institute of Medical Biology CAMS vaccine` = ifelse(str_detect(vaccines, "IMBCAMS"), 1, 0),
    `Janssen Ad26.COV2.S` = ifelse(str_detect(vaccines, "Johnson&Johnson"), 1, 0),
    `Shenzhen Kangtai KCONVAC` = ifelse(str_detect(vaccines, "KCONVAC"), 1, 0),
    `Medicago CoVLP` = ifelse(str_detect(vaccines, "Medicago"), 1, 0),
    `Medigen MVC-COV1901` = ifelse(str_detect(vaccines, "Medigen"), 1, 0),
    `Moderna mRNA-1273` = ifelse(str_detect(vaccines, "Moderna"), 1, 0),
    `Novavax NVX-CoV2373` = ifelse(str_detect(vaccines, "Novavax"), 1, 0),
    `Oxford/AstraZeneca ChAdOx1-S` = ifelse(str_detect(vaccines, "Oxford/AstraZeneca") | str_detect(vaccines, "Covishield"), 1, 0),
    `BioNTech/Pfizer BNT162b2` = ifelse(str_detect(vaccines, "Pfizer/BioNTech"), 1, 0),
    `RIBSP Kazakhstan QazCovid-in` = ifelse(str_detect(vaccines, "QazVac"), 1, 0),
    `Beijing/Sinopharm BBIBP-CorV` = ifelse(str_detect(vaccines, "Sinopharm/Beijing") | str_detect(vaccines, "Sinopharm/HayatVax") | str_detect(vaccines, "BBIBP-CorV"), 1, 0),
    `Wuhan/Sinopharm vaccine` = ifelse(str_detect(vaccines, "Sinopharm/Wuhan"), 1, 0),
    `Sinovac CoronaVac` = ifelse(str_detect(vaccines, "Sinovac"), 1, 0),
    `Instituto Finlay de Vacunas Soberana 02` = ifelse(str_detect(vaccines, "Soberana02"), 1, 0),
    `Instituto Finlay de Vacunas Soberana 01` = ifelse(str_detect(vaccines, "Soberana Plus"), 1, 0),
    `Gamaleya Gam-COVID-Vac/Sputnik V` = ifelse(str_detect(vaccines, "Sputnik"), 1, 0),
    `Erciyes TURKOVAC` = ifelse(str_detect(vaccines, "Turkovac"), 1, 0),
    `AZLB ZF2001` = ifelse(str_detect(vaccines, "RBD-Dimer") | str_detect(vaccines, "ZF2001"), 1, 0),
    `Sanofi/GSK CoV2 preS dTM` = ifelse(str_detect(vaccines, "Sanofi/GSK"), 1, 0),
    `Valneva VLA2001` = ifelse(str_detect(vaccines, "Valneva"), 1, 0)
  ) %>%
  # Create col for each vacc (1/0)
  mutate(
   mRNA = ifelse(`BioNTech/Pfizer BNT162b2`==1 | `Moderna mRNA-1273`==1,1,0),
    Vectored = ifelse(`Oxford/AstraZeneca ChAdOx1-S`==1 | `Janssen Ad26.COV2.S`==1 | `Gamaleya Gam-COVID-Vac/Sputnik V`==1 | `Cansino Ad5-nCoV`==1,1,0),
    Inactivated = ifelse(`Beijing/Sinopharm BBIBP-CorV`==1 | `Sinovac CoronaVac`==1 | `Bharat Covaxin/BBV152`==1 | `Wuhan/Sinopharm vaccine`==1 | `RIBSP Kazakhstan QazCovid-in`==1 | 
                           `Shifa Pharmed COVIran`==1 | `Erciyes TURKOVAC`==1 | `Shenzhen Kangtai KCONVAC`==1 | `Institute of Medical Biology CAMS vaccine`==1 | `Chumakov Center CoviVac`==1 | `Valneva VLA2001`==1,1,0),
    `Protein` = ifelse(`Novavax NVX-CoV2373`==1 | `CIGB CIGB-66/Abdala`==1 | `Instituto Finlay de Vacunas Soberana 02`==1 | `AZLB ZF2001`==1 | `Vector Institute EpiVacCorona`==1 | 
                             `Medigen MVC-COV1901`==1 | `Instituto Finlay de Vacunas Soberana 01`==1 | `Biological E Corbevax`==1 | `Sanofi/GSK CoV2 preS dTM`==1 | `Medicago CoVLP`==1,1,0),
  )

# Save output: data inequalities OWID/WHO vaccine data + metrics + platforms
write.csv(Ineq_vaccine_coverage_platform, "Data/Processed/Ineq_merged_with_vaccine_data.csv")

```


#######################################################################################################
##Fig 4. Vaccine coverage in high-risk populations (health care workers and older adults) and overall.#
#######################################################################################################


```{r}
#High priority metrics
high_priority_metrics <- WHO_hcw_old_data %>% 
  arrange(date, iso) %>% 
  group_by(iso) %>% 
  mutate(
    vac_intro_hcw = ifelse(row_number() == which(cov_hcw_cps > 0)[1], as.character(date), NA),
    vac_F40_hcw = ifelse(row_number() == which(cov_hcw_cps > 0.3999)[1], as.character(date), NA),
    vac_F70_hcw = ifelse(row_number() == which(cov_hcw_cps > 0.6999)[1], as.character(date), NA),
    vac_intro_old = ifelse(row_number() == which(cov_old_cps > 0)[1], as.character(date), NA),
    vac_F40_old = ifelse(row_number() == which(cov_old_cps > 0.3999)[1], as.character(date), NA),
    vac_F70_old = ifelse(row_number() == which(cov_old_cps > 0.6999)[1], as.character(date), NA)
  ) %>%
  filter(!is.na(vac_F70_old) | !is.na(vac_F40_old) | !is.na(vac_intro_old) | !is.na(vac_F70_hcw) | !is.na(vac_F40_hcw) | !is.na(vac_intro_hcw)) %>% 
  select(!(iso:pol_age_old_source)) %>% 
  fill(vac_intro_hcw:vac_F70_old)%>%
  slice_tail() %>%  
  ungroup() %>% 
  mutate(
    V1_hcw_dummy = case_when(
      !is.na(vac_intro_hcw) & ymd(vac_intro_hcw) < ymd("2022-01-01") ~ "yes",
      is.na(vac_intro_hcw) ~ NA,
      TRUE ~ "no"
    ),
    V40_hcw_dummy = case_when(
      !is.na(vac_F40_hcw) & ymd(vac_F40_hcw) < ymd("2022-01-01") ~ "yes",
      is.na(vac_intro_hcw) ~ NA,
      TRUE ~ "no"
    ),
    V70_hcw_dummy = case_when(
      !is.na(vac_F70_hcw) & ymd(vac_F70_hcw) < ymd("2022-07-01") ~ "yes",
      is.na(vac_intro_hcw) ~ NA,
      TRUE ~ "no"
    ),
    V1_old_dummy = case_when(
      !is.na(vac_intro_old) & ymd(vac_intro_old) < ymd("2022-01-01") ~ "yes",
      is.na(vac_intro_old) ~ NA,
      TRUE ~ "no"
    ),
    V40_old_dummy = case_when(
      !is.na(vac_F40_old) & ymd(vac_F40_old) < ymd("2022-01-01") ~ "yes",
      is.na(vac_intro_old) ~ NA,
      TRUE ~ "no"
    ),
    V70_old_dummy = case_when(
      !is.na(vac_F70_old) & ymd(vac_F70_old) < ymd("2022-07-01") ~ "yes",
      is.na(vac_intro_old) ~ NA,
      TRUE ~ "no"
    ),
    V1_hcw_dummy_nd = case_when(
      !is.na(vac_intro_hcw) & ymd(vac_intro_hcw) < ymd("2024-01-01") ~ "yes",
      is.na(vac_intro_hcw) ~ NA,
      TRUE ~ "no"
    ),
    V40_hcw_dummy_nd = case_when(
      !is.na(vac_F40_hcw) & ymd(vac_F40_hcw) < ymd("2024-01-01") ~ "yes",
      is.na(vac_intro_hcw) ~ NA,
      TRUE ~ "no"
    ),
    V70_hcw_dummy_nd = case_when(
      !is.na(vac_F70_hcw) & ymd(vac_F70_hcw) < ymd("2024-01-01") ~ "yes",
      is.na(vac_intro_hcw) ~ NA,
      TRUE ~ "no"
    ),
    V1_old_dummy_nd = case_when(
      !is.na(vac_intro_old) & ymd(vac_intro_old) < ymd("2024-01-01") ~ "yes",
      is.na(vac_intro_old) ~ NA,
      TRUE ~ "no"
    ),
    V40_old_dummy_nd = case_when(
      !is.na(vac_F40_old) & ymd(vac_F40_old) < ymd("2024-01-01") ~ "yes",
      is.na(vac_intro_old) ~ NA,
      TRUE ~ "no"
    ),
    V70_old_dummy_nd = case_when(
      !is.na(vac_F70_old) & ymd(vac_F70_old) < ymd("2024-01-01") ~ "yes",
      is.na(vac_intro_old) ~ NA,
      TRUE ~ "no"
    )
  )

coverage_high_2021_2023<- WHO_hcw_old_data %>% 
  select(iso, date, cov_hcw_cps, cov_old_cps, cov_tot_cps, pop, pop_hcw, pop_old, pol_age_old, pol_age_old_source) %>% 
  filter(date=="2021-12-01" | date=="2023-12-01" | date=="2023-11-01") %>% 
  pivot_wider(., names_from = date, values_from = c(cov_hcw_cps, cov_old_cps, cov_tot_cps, pop, pop_hcw, pop_old, pol_age_old, pol_age_old_source))

Ineq_vaccine_coverage<- left_join(Ineq_vaccine_coverage, high_priority_metrics, by=c("iso3c"= "iso"))

Ineq_vaccine_coverage<- left_join(Ineq_vaccine_coverage, coverage_high_2021_2023, by=c("iso3c"= "iso"))

# Since some countries end reporting at 11/2023 rather than 12/2023, take maximum value 
Ineq_vaccine_coverage <- Ineq_vaccine_coverage %>%
  mutate(
    `cov_hcw_cps_2023-12-01` = if_else(!is.na(`cov_hcw_cps_2023-12-01`), `cov_hcw_cps_2023-12-01`, `cov_hcw_cps_2023-11-01`),
    `cov_old_cps_2023-12-01` = if_else(!is.na(`cov_old_cps_2023-12-01`), `cov_old_cps_2023-12-01`, `cov_old_cps_2023-11-01`),
    `cov_tot_cps_2023-12-01` = if_else(!is.na(`cov_tot_cps_2023-12-01`), `cov_tot_cps_2023-12-01`, `cov_tot_cps_2023-11-01`)
  ) %>%
  select(-ends_with('2023-11-01'))
  
# Save output: data inequalities OWID/WHO vaccine data + metrics + priority groups
write.csv(Ineq_vaccine_coverage, "Data/Processed/Ineq_merged_with_vaccine_data_plus_priority_groups.csv")
```


```{r}
#### Boxplot milestone December 2021/2023 priority
collated = rbind(
  Ineq_vaccine_coverage %>% mutate(coverage = `cov_hcw_cps_2021-12-01`, date = "December 2021", group = "HCWs"),
  Ineq_vaccine_coverage %>% mutate(coverage = `cov_hcw_cps_2023-12-01`, date = "December 2023", group = "HCWs"),
  Ineq_vaccine_coverage %>% mutate(coverage = `cov_old_cps_2021-12-01`, date = "December 2021", group = "Older adults"),
  Ineq_vaccine_coverage %>% mutate(coverage = `cov_old_cps_2023-12-01`, date = "December 2023", group = "Older adults"),
  Ineq_vaccine_coverage %>% mutate(coverage = `cov_tot_cps_2021-12-01`, date = "December 2021", group = "Whole population"),
  Ineq_vaccine_coverage %>% mutate(coverage = `cov_tot_cps_2023-12-01`, date = "December 2023", group = "Whole population")
)

collated$group = factor(collated$group, levels=unique(collated$group))

#Figure 4
# 1) Labels: one count per quartile x date x group
labels_df <- collated %>%
  filter(!is.na(coverage)) %>%
  mutate(q_GNI_per_c_WB = factor(q_GNI_per_c_WB, levels = c(1,2,3,4))) %>%
  group_by(q_GNI_per_c_WB, date, group) %>%
  summarise(n_points = n(), .groups = "drop")

# Set upper limit at 1 for plotting
collated$coverage[collated$coverage>1] = 1 

# 2) Plot
F4_priority_groups <- ggplot(collated) +
  geom_boxplot(aes(x = coverage*100, y = as.factor(q_GNI_per_c_WB)), outlier.shape = NA) +
  geom_point(
    aes(
      x = coverage*100, 
      y = as.factor(q_GNI_per_c_WB),
      color = as.factor(q_GNI_per_c_WB)
    ),
    alpha = 0.5,
    position = position_jitterdodge(dodge.width = 0.5, jitter.width = 1)
  ) +
    geom_text(
    data = labels_df,
    aes(
      y = as.factor(q_GNI_per_c_WB),
      label = paste0("n=", n_points)
    ),
    x = 110,           # place just beyond the 100% tick
    hjust = 0,
    size = 3,
    colour = "grey40", 
    inherit.aes = FALSE
  ) +
  scale_color_manual(values = c("#77AADD","#44BB99","#AAAA00","#EE8866")) +
  scale_x_continuous(breaks = seq(0, 100, 20)) +
  facet_grid(group ~ date) +
  labs(#title = "Health Care workers",
    x = "Primary series coverage (%)", 
    y = "GNI quartile"
  ) +
  guides(color = "none") +
  scale_y_discrete(labels = c("1 (lower\nincome)", "2", "3", "4 (higher\nincome)")) +
  coord_cartesian(xlim = c(0, 115), clip = "off") +  
  theme_minimal() +
  theme(
    panel.spacing = unit(2.0, "lines"),
    legend.position = "bottom",
    legend.text = element_text(size = 8),
    legend.title = element_text(size = 9),
    plot.title = element_text(size = 10),
    plot.margin = margin(5.5, 40, 5.5, 5.5)
  )

F4_priority_groups

# Info for manuscript
n_by_gni_date_group <- collated %>%
  filter(!is.na(coverage)) %>%
  mutate(q_GNI_per_c_WB = factor(q_GNI_per_c_WB, levels = c(1,2,3,4))) %>%
  group_by(q_GNI_per_c_WB, date, group) %>%
  summarise(n_countries = n_distinct(Country.Name), .groups = "drop") %>%
  complete(q_GNI_per_c_WB, date, group, fill = list(n_countries = 0)) %>%
  arrange(group, date, q_GNI_per_c_WB)

n_by_gni_date_group
```



###################################################################################################
# Supplementary Fig. 1. Percentage of countries using different vaccine products by income status.#
###################################################################################################


```{r Supplementary Figure 1}
# Supplementary Fig. 1. Percentage of countries using different vaccine products by income status.
#Pivot longer + plat classification
owid_vaccine_country_g <- Ineq_vaccine_coverage_platform %>% 
    pivot_longer(
    cols = `CIGB CIGB-66/Abdala`: `Valneva VLA2001`,
    names_to = "vaccine",
    values_to = "check",
    values_drop_na = TRUE) %>% 
  filter(
    check== 1
  ) %>% 
  select(-check) %>%
  mutate(
    vac_type = case_when(
      vaccine %in% c("BioNTech/Pfizer BNT162b2", "Moderna mRNA-1273") ~ "mRNA",
      vaccine %in% c("Oxford/AstraZeneca ChAdOx1-S", "Janssen Ad26.COV2.S", "Gamaleya Gam-COVID-Vac/Sputnik V", "Cansino Ad5-nCoV") ~ "Vectored",
      vaccine %in% c("Beijing/Sinopharm BBIBP-CorV", "Sinovac CoronaVac", "Bharat Covaxin/BBV152", "Wuhan/Sinopharm vaccine", "RIBSP Kazakhstan QazCovid-in", "Shifa Pharmed COVIran", "Erciyes TURKOVAC", "Shenzhen Kangtai KCONVAC", "Institute of Medical Biology CAMS vaccine", "Chumakov Center CoviVac", "Valneva VLA2001") ~ "Inactivated",
      vaccine %in% c("Novavax NVX-CoV2373", "CIGB CIGB-66/Abdala", "Instituto Finlay de Vacunas Soberana 02", "AZLB ZF2001", "Vector Institute EpiVacCorona", "Medigen MVC-COV1901", "Instituto Finlay de Vacunas Soberana 01", "Biological E Corbevax", "Sanofi/GSK CoV2 preS dTM", "Medicago CoVLP") ~ "Protein",
      TRUE ~ "Unknown"
    )
  ) 

unique(owid_vaccine_country_g$vaccine)
#Data preparation
vac_prod_distrib_GNI <-owid_vaccine_country_g %>% 
  select(
    Country.Code, q_GNI_per_c_WB, vac_type, vaccine
  ) %>% 
  # group_by(vaccine) %>% 
  # mutate(tot_vac = n()) %>%  
  distinct() %>% 
  filter(
    !is.na(q_GNI_per_c_WB)
  ) %>% 
  group_by(q_GNI_per_c_WB, vaccine, vac_type) %>% 
  summarise(total=n()) %>% 
  # filter(
  #   vac_type=="Inactivated" |vac_type=="mRNA" |vac_type=="Vectored"
  # ) %>%
  mutate(
    countries_quartile=case_when(
      q_GNI_per_c_WB == 1 ~ 50,
      q_GNI_per_c_WB == 2 ~ 51,
      q_GNI_per_c_WB == 3 ~ 51,
      q_GNI_per_c_WB == 4 ~ 51
    )
  ) %>% 
  mutate(
    vac_name_per = round(total / countries_quartile * 100, 1),
    ) 

vac_prod_distrib_GNI

vac_prod_distrib_GNI$vac_type <- factor(vac_prod_distrib_GNI$vac_type, levels = c("mRNA", "Vectored", "Inactivated", "Protein"))


vac_prod_distrib_GNI <- vac_prod_distrib_GNI %>%
  arrange(vac_type, -vac_name_per) %>%
  mutate(order_variable = row_number()) %>%
  ungroup()

# set vaccine factor order
vac_prod_distrib_GNI$vaccine = factor(vac_prod_distrib_GNI$vaccine, levels = unique(vac_prod_distrib_GNI$vaccine))


#Plot Extended data 1
ED_F1_vac_prod_distrib_GNI<- ggplot(vac_prod_distrib_GNI, aes(x = as.factor(q_GNI_per_c_WB), y = vac_name_per, fill = as.factor(vac_type))) +
  geom_bar(stat = "identity", position = "dodge") +
  coord_flip()+
  facet_wrap(~vaccine,ncol = 5 ) +
  labs(title = "",
       x = "GNI quartile",
       y = " % countries using vaccine",
       fill = "Vaccine Type") +
  scale_fill_manual(values = c("#F0E442", "#CC79A7","#D55E00","#0072B2" ),
    labels = c("mRNA","Vectored", "Inactivated", "Protein")) +
  theme_minimal() +
  scale_x_discrete(labels = c("1 (lower\nincome)", "2", "3", "4 (higher\nincome)")) + 
  theme(
    legend.position = "bottom",
    legend.text = element_text(size = 8),  # Adjust text size
    legend.title = element_text(size = 9),  # Adjust legend title size
    plot.title = element_text(size = 10) 
  )
  
ED_F1_vac_prod_distrib_GNI

```



########################################################################################
# Supplementary Fig. 7. Map of countries by per capita Gross National Income quartile.###
########################################################################################


```{r}
# Supplementary Fig. 7. Map of countries by per capita Gross National Income quartile.
#Geographical information
countries <-  ne_countries(returnclass = "sf") %>% 
  mutate(iso3c = countrycode(admin, origin = "country.name", destination = "iso3c")) %>% 
  filter(iso3c!="ATA")
  

world_ineq_gral <- 
  merge(countries, Ineq_vaccine_coverage, by.x = "iso3c", by.y = "Country.Code", all = TRUE)

world_ineq_gral <- st_as_sf(world_ineq_gral)
#Plot Map
ED_F8_map_GNI_q <- ggplot(data = world_ineq_gral) +
  geom_sf(aes(fill = as.factor(q_GNI_per_c_WB)), position = "identity") + 
  scale_fill_manual(values = c("#77AADD","#44BB99","#AAAA00","#EE8866"),
                    labels = c("1 (lower\nincome)", "2", "3", "4 (higher\nincome)","no data\navailable")) +
  labs(
    fill = "GNI quartile") +
  theme_void()+
  guides(color = guide_legend(title.position = "bottom", nrow = 1))+
  theme(
    legend.position = "bottom",
    legend.text = element_text(size = 10),  # Adjust text size
    legend.title = element_text(size = 10),  # Adjust legend title size
    plot.title = element_text(size = 10) 
  )

ED_F8_map_GNI_q


```


##############################
#####TABLES AND METRICS#######
##############################

########################################################################################
#Supplementary Table 1. Vaccine introduction and scale-up milestones by income status.#
########################################################################################
# Summary vaccine introduction and scale-up

```{r}
# Summary milestones
summary_milestone_achieved <- Ineq_vaccine_coverage %>% 
  group_by(q_GNI_per_c_WB) %>% 
  summarise(
    V1 = sum(V1_dummy == "yes"),
    V40 = sum(V40_dummy == "yes"),
    V70 = sum(V70_dummy == "yes"),
    V1_nd = sum(V1_dummy_no_date == "yes"),
    V40_nd = sum(V40_dummy_no_date == "yes"),
    V70_nd = sum(V70_dummy_no_date == "yes"),
    total = n(),
    V1_per = round(sum(V1_dummy == "yes") / n() * 100, 1),
    V40_per = round(sum(V40_dummy == "yes") / V1 * 100, 1),
    V70_per = round(sum(V70_dummy == "yes") / V1 * 100, 1),
    V1_per_nd = round(sum(V1_dummy_no_date == "yes") / n() * 100, 1),
    V40_per_nd = round(sum(V40_dummy_no_date == "yes") / V1_nd * 100, 1),
    V70_per_nd = round(sum(V70_dummy_no_date == "yes") / V1_nd * 100, 1),
  )

summary_milestone_achieved

#Percentage milestone with and without considering date of achievement
table_milestone_achieved <- summary_milestone_achieved %>% 
  mutate(
    intro_d_ach = paste0(V1, "/51", " (", V1_per, "%)"),
    intro_nd_ach = paste0(V1_nd, "/51", " (", V1_per_nd, "%)"),
    scale_40_d_ach = paste0(V40, "/", V1_nd, " (", V40_per, "%)"),
    scale_40_nd_ach = paste0(V40_nd, "/", V1_nd, " (", V40_per_nd, "%)"),
    scale_70_d_ach = paste0(V70, "/", V1_nd, " (", V70_per, "%)"),
    scale_70_nd_ach = paste0(V70_nd, "/", V1_nd, " (", V70_per_nd, "%)")    
  ) %>% 
  select( q_GNI_per_c_WB, intro_d_ach:scale_70_nd_ach)

#Time in days and date achived median without considering date of achievement
summary_vacc_intro_scale_nd <- Ineq_vaccine_coverage %>%
  group_by(q_GNI_per_c_WB) %>%
  summarize(
    # Summary statistics for scale-up delays to reach 40% coverage (in days)
    median_value_40 = round(median(diff_intro_F40, na.rm = TRUE), 0),
    q25_40 = round(quantile(diff_intro_F40, 0.25, na.rm = TRUE), 0),
    q75_40 = round(quantile(diff_intro_F40, 0.75, na.rm = TRUE), 0),
    # Summary statistics for scale-up delays to reach 70% coverage (in days)
    median_value_70 = round(median(diff_intro_F70, na.rm = TRUE), 0),
    q25_70 = round(quantile(diff_intro_F70, 0.25, na.rm = TRUE), 0),
    q75_70 = round(quantile(diff_intro_F70, 0.75, na.rm = TRUE), 0),
    # Dates for vaccine introduction and scale-ups (median and quartiles)
    median_value_1_date = as.Date(median(as.numeric(min_vac_intro), na.rm = TRUE), origin = "1970-01-01"),
    q25_1_date = as.Date(quantile(as.numeric(min_vac_intro), 0.25, na.rm = TRUE), origin = "1970-01-01"),
    q75_1_date = as.Date(quantile(as.numeric(min_vac_intro), 0.75, na.rm = TRUE), origin = "1970-01-01"),
    # Dates for reaching 40% coverage
    median_value_40_date = as.Date(median(as.numeric(as.Date(owid_vac_F40)), na.rm = TRUE), origin = "1970-01-01"),
    q25_40_date = as.Date(quantile(as.numeric(as.Date(owid_vac_F40)), 0.25, na.rm = TRUE), origin = "1970-01-01"),
    q75_40_date = as.Date(quantile(as.numeric(as.Date(owid_vac_F40)), 0.75, na.rm = TRUE), origin = "1970-01-01"),
    # Dates for reaching 70% coverage
    median_value_70_date = as.Date(median(as.numeric(as.Date(owid_vac_F70)), na.rm = TRUE), origin = "1970-01-01"),
    q25_70_date = as.Date(quantile(as.numeric(as.Date(owid_vac_F70)), 0.25, na.rm = TRUE), origin = "1970-01-01"),
    q75_70_date = as.Date(quantile(as.numeric(as.Date(owid_vac_F70)), 0.75, na.rm = TRUE), origin = "1970-01-01")
  ) %>%
  mutate(
    # Create readable summary columns for delays
    scale_40 = paste0(median_value_40, " (", q25_40, "-", q75_40, ")"),
    scale_70 = paste0(median_value_70, " (", q25_70, "-", q75_70, ")"),
    # Create readable summary columns for dates
    introduction_date = paste0(format(median_value_1_date, "%d/%m/%Y"), " (", 
                                format(q25_1_date, "%d/%m/%Y"), "-", 
                                format(q75_1_date, "%d/%m/%Y"), ")"),
    scale_40_date = paste0(format(median_value_40_date, "%d/%m/%Y"), " (", 
                           format(q25_40_date, "%d/%m/%Y"), "-", 
                           format(q75_40_date, "%d/%m/%Y"), ")"),
    scale_70_date = paste0(format(median_value_70_date, "%d/%m/%Y"), " (", 
                           format(q25_70_date, "%d/%m/%Y"), "-", 
                           format(q75_70_date, "%d/%m/%Y"), ")")
  ) %>%
  # Select and organize the relevant output columns
  select(
    q_GNI_per_c_WB, 
    scale_40, scale_70, 
    introduction_date, scale_40_date, scale_70_date
  )

summary_vacc_intro_scale_nd


#Time in days and date achieved median without considering date of achievement
summary_vacc_intro_scale_d <- Ineq_vaccine_coverage %>%
  group_by(q_GNI_per_c_WB) %>%
  summarize(
    # Introduction (subset by V1_dummy == "yes")
    median_value_1_date_d = as.Date(median(as.numeric(as.Date(min_vac_intro[V1_dummy == "yes"])), na.rm = TRUE), origin = "1970-01-01"),
    q25_1_date_d = as.Date(quantile(as.numeric(as.Date(min_vac_intro[V1_dummy == "yes"])), 0.25, na.rm = TRUE), origin = "1970-01-01"),
    q75_1_date_d = as.Date(quantile(as.numeric(as.Date(min_vac_intro[V1_dummy == "yes"])), 0.75, na.rm = TRUE), origin = "1970-01-01"),

    # 40% coverage (subset by V40_dummy == "yes")
    median_value_40_d = round(median(diff_intro_F40[V40_dummy == "yes"], na.rm = TRUE), 0),
    q25_40_d = round(quantile(diff_intro_F40[V40_dummy == "yes"], 0.25, na.rm = TRUE), 0),
    q75_40_d = round(quantile(diff_intro_F40[V40_dummy == "yes"], 0.75, na.rm = TRUE), 0),
    median_value_40_date_d = as.Date(median(as.numeric(as.Date(owid_vac_F40[V40_dummy == "yes"])), na.rm = TRUE), origin = "1970-01-01"),
    q25_40_date_d = as.Date(quantile(as.numeric(as.Date(owid_vac_F40[V40_dummy == "yes"])), 0.25, na.rm = TRUE), origin = "1970-01-01"),
    q75_40_date_d = as.Date(quantile(as.numeric(as.Date(owid_vac_F40[V40_dummy == "yes"])), 0.75, na.rm = TRUE), origin = "1970-01-01"),

    # 70% coverage (subset by V70_dummy == "yes")
    median_value_70_d = round(median(diff_intro_F70[V70_dummy == "yes"], na.rm = TRUE), 0),
    q25_70_d = round(quantile(diff_intro_F70[V70_dummy == "yes"], 0.25, na.rm = TRUE), 0),
    q75_70_d = round(quantile(diff_intro_F70[V70_dummy == "yes"], 0.75, na.rm = TRUE), 0),
    median_value_70_date_d = as.Date(median(as.numeric(as.Date(owid_vac_F70[V70_dummy == "yes"])), na.rm = TRUE), origin = "1970-01-01"),
    q25_70_date_d = as.Date(quantile(as.numeric(as.Date(owid_vac_F70[V70_dummy == "yes"])), 0.25, na.rm = TRUE), origin = "1970-01-01"),
    q75_70_date_d = as.Date(quantile(as.numeric(as.Date(owid_vac_F70[V70_dummy == "yes"])), 0.75, na.rm = TRUE), origin = "1970-01-01")
  ) %>%
  mutate(
    scale_40_d = paste0(median_value_40_d, " (", q25_40_d, "-", q75_40_d, ")"),
    scale_70_d = paste0(median_value_70_d, " (", q25_70_d, "-", q75_70_d, ")"),

    introduction_date_d = paste0(format(median_value_1_date_d, "%d/%m/%Y"), " (", 
                                 format(q25_1_date_d, "%d/%m/%Y"), "-", 
                                 format(q75_1_date_d, "%d/%m/%Y"), ")"),
    scale_40_date_d = paste0(format(median_value_40_date_d, "%d/%m/%Y"), " (", 
                             format(q25_40_date_d, "%d/%m/%Y"), "-", 
                             format(q75_40_date_d, "%d/%m/%Y"), ")"),
    scale_70_date_d = paste0(format(median_value_70_date_d, "%d/%m/%Y"), " (", 
                             format(q25_70_date_d, "%d/%m/%Y"), "-", 
                             format(q75_70_date_d, "%d/%m/%Y"), ")")
  ) %>%
  select(
    q_GNI_per_c_WB, 
    introduction_date_d,
    scale_40_d, scale_40_date_d,
    scale_70_d, scale_70_date_d
  )
summary_vacc_intro_scale_d


# Collated table
full_milestone_summary <- table_milestone_achieved %>%
  left_join(summary_vacc_intro_scale_nd, by = "q_GNI_per_c_WB") %>%
  left_join(summary_vacc_intro_scale_d, by = "q_GNI_per_c_WB")

full_milestone_summary_flip <-   full_milestone_summary %>% 
    pivot_longer(
    cols = -q_GNI_per_c_WB,
    names_to = "indicator",
    values_to = "value"
  ) %>% 
    pivot_wider(
    names_from = q_GNI_per_c_WB,
    values_from = value,
    names_prefix = "GNI_"
  )

full_milestone_summary_flip_ordered <- full_milestone_summary_flip %>%
  filter(indicator %in% c(
    "intro_d_ach",
    "intro_nd_ach",
    "introduction",
    "introduction_date",
    "scale_40_d_ach",
    "scale_40_d",
    "scale_40_date_d",
    "scale_40_nd_ach",
    "scale_40",
    "scale_40_date",
    "scale_70_d_ach",
    "scale_70_d",
    "scale_70_date_d",
    "scale_70_nd_ach",
    "scale_70",
    "scale_70_date"
  )) %>%
  mutate(indicator = factor(indicator, levels = c(
    "intro_d_ach",
    "intro_nd_ach",
    "introduction",
    "introduction_date",
    "scale_40_d_ach",
    "scale_40_d",
    "scale_40_date_d",
    "scale_40_nd_ach",
    "scale_40",
    "scale_40_date",
    "scale_70_d_ach",
    "scale_70_d",
    "scale_70_date_d",
    "scale_70_nd_ach",
    "scale_70",
    "scale_70_date"
  ))) %>%
  arrange(indicator)
full_milestone_summary_flip_ordered

```



#################################################################################################################
#Supplementary Table 2. Vaccine coverage in high-risk populations and overall according to WHO/UNICEF database..#
#################################################################################################################
```{r}
#Overall coverage dec 2021/ 2023
summary_overall_cov_21_23 <- collated %>%
  select(iso3c, q_GNI_per_c_WB, perc_HCW, perc_old, `cov_hcw_cps_2023-12-01`:`cov_tot_cps_2021-12-01`) %>% 
  unique() %>% 
  filter(!is.na(iso3c)) %>% 
  group_by(q_GNI_per_c_WB) %>%
  summarize(
    n = length(q_GNI_per_c_WB),
    n_hcw_dec21 = sum(!is.na(`cov_hcw_cps_2021-12-01`)),
    n_hcw_dec23 = sum(!is.na(`cov_hcw_cps_2023-12-01`)),
    n_old_dec21 = sum(!is.na(`cov_old_cps_2021-12-01`)),
    n_old_dec23 = sum(!is.na(`cov_old_cps_2023-12-01`)),
    n_tot_dec21 = sum(!is.na(`cov_tot_cps_2021-12-01`)),
    n_tot_dec23 = sum(!is.na(`cov_tot_cps_2023-12-01`)),
    median_hcw_perc = round((median(as.numeric(`perc_HCW`), na.rm = TRUE)*100),1),
    q25_hcw_perc = round((quantile(as.numeric(`perc_HCW`), 0.25, na.rm = TRUE)*100),1),
    q75_hcw_perc = round((quantile(as.numeric(`perc_HCW`), 0.75, na.rm = TRUE)*100),1),
    median_old_perc = round((median(as.numeric(`perc_old`), na.rm = TRUE)*100),0),
    q25_old_perc = round((quantile(as.numeric(`perc_old`), 0.25, na.rm = TRUE)*100),0),
    q75_old_perc = round((quantile(as.numeric(`perc_old`), 0.75, na.rm = TRUE)*100),0),
    median_hcw_dec21 = round((median(as.numeric(`cov_hcw_cps_2021-12-01`), na.rm = TRUE)*100),0),
    q25_hcw_dec21 = round((quantile(as.numeric(`cov_hcw_cps_2021-12-01`), 0.25, na.rm = TRUE)*100),0),
    q75_hcw_dec21 = round((quantile(as.numeric(`cov_hcw_cps_2021-12-01`), 0.75, na.rm = TRUE)*100),0),
    median_hcw_dec23 = round((median(as.numeric(`cov_hcw_cps_2023-12-01`), na.rm = TRUE)*100),0),
    q25_hcw_dec23 = round((quantile(as.numeric(`cov_hcw_cps_2023-12-01`), 0.25, na.rm = TRUE)*100),0),
    q75_hcw_dec23 = round((quantile(as.numeric(`cov_hcw_cps_2023-12-01`), 0.75, na.rm = TRUE)*100),0),
    median_old_dec21 = round((median(as.numeric(`cov_old_cps_2021-12-01`), na.rm = TRUE)*100),0),
    q25_old_dec21 = round((quantile(as.numeric(`cov_old_cps_2021-12-01`), 0.25, na.rm = TRUE)*100),0),
    q75_old_dec21 = round((quantile(as.numeric(`cov_old_cps_2021-12-01`), 0.75, na.rm = TRUE)*100),0),
    median_old_dec23 = round((median(as.numeric(`cov_old_cps_2023-12-01`), na.rm = TRUE)*100),0),
    q25_old_dec23 = round((quantile(as.numeric(`cov_old_cps_2023-12-01`), 0.25, na.rm = TRUE)*100),0),
    q75_old_dec23 = round((quantile(as.numeric(`cov_old_cps_2023-12-01`), 0.75, na.rm = TRUE)*100),0),
    median_tot_dec21 = round((median(as.numeric(`cov_tot_cps_2021-12-01`), na.rm = TRUE)*100),0),
    q25_tot_dec21 = round((quantile(as.numeric(`cov_tot_cps_2021-12-01`), 0.25, na.rm = TRUE)*100),0),
    q75_tot_dec21 = round((quantile(as.numeric(`cov_tot_cps_2021-12-01`), 0.75, na.rm = TRUE)*100),0),
    median_tot_dec23 = round((median(as.numeric(`cov_tot_cps_2023-12-01`), na.rm = TRUE)*100),0),
    q25_tot_dec23 = round((quantile(as.numeric(`cov_tot_cps_2023-12-01`), 0.25, na.rm = TRUE)*100),0),
    q75_tot_dec23 = round((quantile(as.numeric(`cov_tot_cps_2023-12-01`), 0.75, na.rm = TRUE)*100),0)
  ) %>% 
  mutate(
    hcw_perc = paste0( median_hcw_perc, "% (", q25_hcw_perc, "%-", q75_hcw_perc, "%)"),
    hcw_dec21 = paste0( median_hcw_dec21, "% (", q25_hcw_dec21, "%-", q75_hcw_dec21, "%)"),
    hcw_dec23 = paste0( median_hcw_dec23, "% (", q25_hcw_dec23, "%-", q75_hcw_dec23, "%)"),
    old_perc = paste0( median_old_perc, "% (", q25_old_perc, "%-", q75_old_perc, "%)"),
    old_dec21 = paste0( median_old_dec21, "% (", q25_old_dec21, "%-", q75_old_dec21, "%)"),
    old_dec23 = paste0( median_old_dec23, "% (", q25_old_dec23, "%-", q75_old_dec23, "%)"),
    tot_dec21 = paste0( median_tot_dec21, "% (", q25_tot_dec21, "%-", q75_tot_dec21, "%)"),
    tot_dec23 = paste0( median_tot_dec23, "% (", q25_tot_dec23, "%-", q75_tot_dec23, "%)")
  ) %>% 
  select(q_GNI_per_c_WB:n_tot_dec23, hcw_perc:tot_dec23) 
summary_overall_cov_21_23
```


###################################################################
#Supplementary Table 3.  Age thresholds for defining older adults.#
###################################################################

## Description of definitions for older age by GNI
```{r}
# Review unique values used for age criteria
unique(Ineq_vaccine_coverage$`pol_age_old_2021-12-01`)
unique(Ineq_vaccine_coverage$`pol_age_old_2023-12-01`)

# Re-assign 0s to NAs
Ineq_vaccine_coverage$`pol_age_old_2021-12-01`[Ineq_vaccine_coverage$`pol_age_old_2021-12-01`==0] = NA
Ineq_vaccine_coverage$`pol_age_old_2023-12-01`[Ineq_vaccine_coverage$`pol_age_old_2023-12-01`==0] = NA

# Re-assign to not reported if no coverage data
Ineq_vaccine_coverage$`pol_age_old_2021-12-01`[is.na(Ineq_vaccine_coverage$`cov_old_cps_2021-12-01`)] = "coverage not reported"
Ineq_vaccine_coverage$`pol_age_old_2023-12-01`[is.na(Ineq_vaccine_coverage$`cov_old_cps_2023-12-01`)] = "coverage not reported"

# Re-assign to 'not defined' if coverage reported without definition
Ineq_vaccine_coverage$`pol_age_old_2021-12-01`[!is.na(Ineq_vaccine_coverage$`cov_old_cps_2021-12-01`) & is.na(Ineq_vaccine_coverage$`pol_age_old_2021-12-01`)] = "coverage reported without definition"
Ineq_vaccine_coverage$`pol_age_old_2023-12-01`[!is.na(Ineq_vaccine_coverage$`cov_old_cps_2023-12-01`) & is.na(Ineq_vaccine_coverage$`pol_age_old_2023-12-01`)] = "coverage reported without definition"

# Tabulate criteria by GNI
table(Ineq_vaccine_coverage$q_GNI_per_c_WB, Ineq_vaccine_coverage$`pol_age_old_2021-12-01`)
table(Ineq_vaccine_coverage$q_GNI_per_c_WB, Ineq_vaccine_coverage$`pol_age_old_2023-12-01`)

# Tabulate N reporting
table(Ineq_vaccine_coverage$q_GNI_per_c_WB, !is.na(Ineq_vaccine_coverage$`cov_hcw_cps_2021-12-01`))
table(Ineq_vaccine_coverage$q_GNI_per_c_WB, !is.na(Ineq_vaccine_coverage$`cov_hcw_cps_2023-12-01`))
table(Ineq_vaccine_coverage$q_GNI_per_c_WB, !is.na(Ineq_vaccine_coverage$`cov_old_cps_2021-12-01`))
table(Ineq_vaccine_coverage$q_GNI_per_c_WB, !is.na(Ineq_vaccine_coverage$`cov_old_cps_2023-12-01`))
```



#########################
#For supplementary database#
########################


```{r}
# Supplementary table Intro+Scale-up
for_sup_table_risk_pop <- Ineq_vaccine_coverage %>%
  select(
    Country.Code, vac_intro_hcw, vac_F40_hcw, vac_F70_hcw, 
    vac_intro_old, vac_F40_old, vac_F70_old, 
    perc_HCW, perc_old, 
    `cov_hcw_cps_2023-12-01`:`pol_age_old_source_2021-12-01`
  )

write.csv(for_sup_table_risk_pop, "Data/Processed/Ineq_merged_with_vaccine_priority_groups.csv")
```

###############
#Other metrics#
###############


```{r}
# HCW

#median dates to built groups
summary_vacc_intro_scale_hcw <- Ineq_vaccine_coverage %>%
  mutate(
    vac_intro_hcw_month = floor_date(as.Date(vac_intro_hcw), "month"),
    vac_F40_hcw_month = floor_date(as.Date(vac_F40_hcw), "month"),
    vac_F70_hcw_month = floor_date(as.Date(vac_F70_hcw), "month")
  ) %>%
  group_by(q_GNI_per_c_WB) %>%
  summarize(
    n = length(q_GNI_per_c_WB),
    n_intro = sum(!is.na(vac_intro_hcw_month)),
    median_value_1 = format(as.Date(median(as.numeric(vac_intro_hcw_month), na.rm = TRUE), origin = "1970-01-01"), "%Y/%m"),
    q25_1 = format(as.Date(quantile(as.numeric(vac_intro_hcw_month), 0.25, na.rm = TRUE), origin = "1970-01-01"), "%Y/%m"),
    q75_1 = format(as.Date(quantile(as.numeric(vac_intro_hcw_month), 0.75, na.rm = TRUE), origin = "1970-01-01"), "%Y/%m"),
    median_value_40 = format(as.Date(median(as.numeric(vac_F40_hcw_month), na.rm = TRUE), origin = "1970-01-01"), "%Y/%m"),
    q25_40 = format(as.Date(quantile(as.numeric(vac_F40_hcw_month), 0.25, na.rm = TRUE), origin = "1970-01-01"), "%Y/%m"),
    q75_40 = format(as.Date(quantile(as.numeric(vac_F40_hcw_month), 0.75, na.rm = TRUE), origin = "1970-01-01"), "%Y/%m"),
    median_value_70 = format(as.Date(median(as.numeric(vac_F70_hcw_month), na.rm = TRUE), origin = "1970-01-01"), "%Y/%m"),
    q25_70 = format(as.Date(quantile(as.numeric(vac_F70_hcw_month), 0.25, na.rm = TRUE), origin = "1970-01-01"), "%Y/%m"),
    q75_70 = format(as.Date(quantile(as.numeric(vac_F70_hcw_month), 0.75, na.rm = TRUE), origin = "1970-01-01"), "%Y/%m")
  ) %>% 
  mutate(
    introduction = paste0( median_value_1, " (", q25_1, "-", q75_1, ")"),
    scale_40 = paste0( median_value_40, " (", q25_40, "-", q75_40, ")"),
    scale_70 = paste0( median_value_70, " (", q25_70, "-", q75_70, ")")
  ) %>% 
  select(q_GNI_per_c_WB, introduction, scale_40, scale_70) 
summary_vacc_intro_scale_hcw

# Milestones
hcw_milestone_achieved <- Ineq_vaccine_coverage %>%
  group_by(q_GNI_per_c_WB) %>%
  summarise(
    V1 = sum(V1_hcw_dummy == "yes", na.rm = TRUE),
    V40 = sum(V40_hcw_dummy == "yes", na.rm = TRUE),
    V70 = sum(V70_hcw_dummy == "yes", na.rm = TRUE),
    V1_nd = sum(V1_hcw_dummy_nd == "yes", na.rm = TRUE),
    V40_nd = sum(V40_hcw_dummy_nd == "yes", na.rm = TRUE),
    V70_nd = sum(V70_hcw_dummy_nd == "yes", na.rm = TRUE),
    total = sum(V1_hcw_dummy_nd == "yes", na.rm = TRUE),
    V1_per = round(sum(V1_hcw_dummy == "yes", na.rm = TRUE) / total * 100, 0),
    V40_per = round(sum(V40_hcw_dummy == "yes", na.rm = TRUE) / total * 100, 0),
    V70_per = round(sum(V70_hcw_dummy == "yes", na.rm = TRUE) / total * 100, 0),
    V1_per_nd = round(sum(V1_hcw_dummy_nd == "yes", na.rm = TRUE) / total * 100, 0),
    V40_per_nd = round(sum(V40_hcw_dummy_nd == "yes", na.rm = TRUE) / total * 100, 0),
    V70_per_nd = round(sum(V70_hcw_dummy_nd == "yes", na.rm = TRUE) / total * 100, 0)
  ) %>% 
  mutate(
    V1_per_number = paste0(V1, "/", total, " (", V1_per, "%)"),
    V40_per_number = paste0(V40, "/", total, " (", V40_per, "%)"),
    V70_per_number = paste0(V70, "/", total, " (", V70_per, "%)"),
    V1_per_nd_number = paste0(V1_nd, "/", total, " (", V1_per_nd, "%)"),
    V40_per_nd_number = paste0(V40_nd, "/", total, " (", V40_per_nd, "%)"),
    V70_per_nd_number = paste0(V70_nd, "/", total, " (", V70_per_nd, "%)")
  ) %>% 
  select(q_GNI_per_c_WB, V1_per_number:V70_per_nd_number)
hcw_milestone_achieved

```




```{r}
# old

#median dates to built groups
summary_vacc_intro_scale_old <- Ineq_vaccine_coverage %>%
  mutate(
    vac_intro_old_month = floor_date(as.Date(vac_intro_old), "month"),
    vac_F40_old_month = floor_date(as.Date(vac_F40_old), "month"),
    vac_F70_old_month = floor_date(as.Date(vac_F70_old), "month")
  ) %>%
  group_by(q_GNI_per_c_WB) %>%
  summarize(
    n = length(q_GNI_per_c_WB),
    n_intro = sum(!is.na(vac_intro_old_month)),
    median_value_1 = format(as.Date(median(as.numeric(vac_intro_old_month), na.rm = TRUE), origin = "1970-01-01"), "%Y/%m"),
    q25_1 = format(as.Date(quantile(as.numeric(vac_intro_old_month), 0.25, na.rm = TRUE), origin = "1970-01-01"), "%Y/%m"),
    q75_1 = format(as.Date(quantile(as.numeric(vac_intro_old_month), 0.75, na.rm = TRUE), origin = "1970-01-01"), "%Y/%m"),
    median_value_40 = format(as.Date(median(as.numeric(vac_F40_old_month), na.rm = TRUE), origin = "1970-01-01"), "%Y/%m"),
    q25_40 = format(as.Date(quantile(as.numeric(vac_F40_old_month), 0.25, na.rm = TRUE), origin = "1970-01-01"), "%Y/%m"),
    q75_40 = format(as.Date(quantile(as.numeric(vac_F40_old_month), 0.75, na.rm = TRUE), origin = "1970-01-01"), "%Y/%m"),
    median_value_70 = format(as.Date(median(as.numeric(vac_F70_old_month), na.rm = TRUE), origin = "1970-01-01"), "%Y/%m"),
    q25_70 = format(as.Date(quantile(as.numeric(vac_F70_old_month), 0.25, na.rm = TRUE), origin = "1970-01-01"), "%Y/%m"),
    q75_70 = format(as.Date(quantile(as.numeric(vac_F70_old_month), 0.75, na.rm = TRUE), origin = "1970-01-01"), "%Y/%m")
  ) %>% 
  mutate(
    introduction = paste0( median_value_1, " (", q25_1, "-", q75_1, ")"),
    scale_40 = paste0( median_value_40, " (", q25_40, "-", q75_40, ")"),
    scale_70 = paste0( median_value_70, " (", q25_70, "-", q75_70, ")")
  ) %>% 
  select(q_GNI_per_c_WB, introduction, scale_40, scale_70) 
summary_vacc_intro_scale_old

# Milestones
old_milestone_achieved <- Ineq_vaccine_coverage %>%
  group_by(q_GNI_per_c_WB) %>%
  summarise(
    V1 = sum(V1_old_dummy == "yes", na.rm = TRUE),
    V40 = sum(V40_old_dummy == "yes", na.rm = TRUE),
    V70 = sum(V70_old_dummy == "yes", na.rm = TRUE),
    V1_nd = sum(V1_old_dummy_nd == "yes", na.rm = TRUE),
    V40_nd = sum(V40_old_dummy_nd == "yes", na.rm = TRUE),
    V70_nd = sum(V70_old_dummy_nd == "yes", na.rm = TRUE),
    total = sum(V1_old_dummy_nd == "yes", na.rm = TRUE),
    V1_per = round(sum(V1_old_dummy == "yes", na.rm = TRUE) / total * 100, 0),
    V40_per = round(sum(V40_old_dummy == "yes", na.rm = TRUE) / total * 100, 0),
    V70_per = round(sum(V70_old_dummy == "yes", na.rm = TRUE) / total * 100, 0),
    V1_per_nd = round(sum(V1_old_dummy_nd == "yes", na.rm = TRUE) / total * 100, 0),
    V40_per_nd = round(sum(V40_old_dummy_nd == "yes", na.rm = TRUE) / total * 100, 0),
    V70_per_nd = round(sum(V70_old_dummy_nd == "yes", na.rm = TRUE) / total * 100, 0)
  ) %>% 
  mutate(
    V1_per_number = paste0(V1, "/", total, " (", V1_per, "%)"),
    V40_per_number = paste0(V40, "/", total, " (", V40_per, "%)"),
    V70_per_number = paste0(V70, "/", total, " (", V70_per, "%)"),
    V1_per_nd_number = paste0(V1_nd, "/", total, " (", V1_per_nd, "%)"),
    V40_per_nd_number = paste0(V40_nd, "/", total, " (", V40_per_nd, "%)"),
    V70_per_nd_number = paste0(V70_nd, "/", total, " (", V70_per_nd, "%)")
  ) %>% 
  select(q_GNI_per_c_WB, V1_per_number:V70_per_nd_number)
old_milestone_achieved

```


# Session info
```{r, session info}
print(sessionInfo())
```
