==> Scripts/4 VE description (Fig 2).Rmd <==
---
title: "4 VE description (Fig 2)"
author: "Martina Pesce & Edward P. K. Parker"
date: "2023-11-30"
output: html_document
editor_options: 
  chunk_output_type: console
---


#### Inputs
- Processed/COVID19VE_ineq_merged.csv: View-Hub VE data merged with country income data
- Processed/Ineq_merged_with_vaccine_data.csv:
- Processed/Ineq_merged_with_vaccine_priority_groups.csv:
- Input/Vaccine_data/owid-covid-data.csv: Our World in Data COVID-19 database
- Input/Vaccine_data/owid_vaccine_plat.csv: Our World in Data COVID-19 vaccine platforms
- Input/Vaccine_data/vaccination-data.csv: WHO COVID-19 vaccine database

#### Outputs
- Fig. 1. Global disparities in COVID-19 vaccine introduction, scale-up, and vaccine effectiveness evaluation.
- Fig. 2. Disparities in COVID-19 vaccine introduction, scale-up and vaccine effectiveness evaluation by income quartile.
- Table 1. Characteristics of primary series vaccine effectiveness studies reporting on country-specific estimates.
- Supplementary Fig. 3. Distribution of vaccine effectiveness papers over time.
- Curated_dataset: Curated_datasets/Ineq_merged_with_vaccine_intro_scale_plat_priority_groups_VE.csv

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```




```{r Libraries}
#Libraries
library(tidyverse)
library(readr)
library(rnaturalearth)
library(sf)
library(lubridate)
library(countrycode)
library(patchwork) 
library(cowplot)
```


```{r Data base}
#Databases

#Eff database
COVID19VE_ineq_merged <- read_csv("Data/Processed/COVID19VE_ineq_merged.csv") 

#Ineq database
Ineq_merged <- read_csv("Data/Processed/Ineq_merged_with_vaccine_data.csv") 

#Vaccine info priority groups
Ineq_merged_prior_group <- read_csv("Data/Processed/Ineq_merged_with_vaccine_priority_groups.csv") %>% 
  select(-`...1`)%>% 
  rename_with(~ paste0("who_unicef_", .x), .cols = -Country.Code)

```

#################################################################################################################
## Fig. 1. Global disparities in COVID-19 vaccine introduction, scale-up, and vaccine effectiveness evaluation.## 
#################################################################################################################

```{r data for map}
#Create table with study counts and introduction dates by country
Country_studies <- COVID19VE_ineq_merged %>%
  group_by(Country.Code, country_clean) %>%
  summarise(
    min_VE_pub_date = min(as.Date(min_pub_date)),
    n_VE_studies = n_distinct(Study_id_f),
    perc_VE_studies = round(n_distinct(Study_id_f)/423*100,0)
  )
nrow(Country_studies) # 48 distinct countries with country-specific VE estimates
nstudies = sum(Country_studies$n_VE_studies) 
nstudies # 423 unique studies
Country_studies %>% arrange(-n_VE_studies) %>% filter(n_VE_studies>10)

# Check all country iso codes represented in inequity data
all(Country_studies$Country.Code %in% Ineq_merged$Country.Code)

#Merge and create study categories
Ineq_merged <- left_join(Ineq_merged, Country_studies, by="Country.Code") %>% 
  mutate(
      study_dummy = case_when(
      !is.na(n_VE_studies) ~ "yes",
      TRUE ~ "no"
    )) %>%
  # add time to publication metrics
   mutate(
    diff_V40_VEpub = as.Date(min_VE_pub_date) - as.Date(owid_vac_F40),
    diff_intro_min_VE_pub = as.Date(min_VE_pub_date) - as.Date(min_vac_intro)
    )

#Geographical information

countries <-  ne_countries(returnclass = "sf") %>% 
  mutate(iso3c = countrycode(admin, origin = "country.name", destination = "iso3c")) %>% 
  filter(iso3c!="ATA")
  

world_ineq_gral <- 
  merge(countries, Ineq_merged, by.x = "iso3c", by.y = "Country.Code", all = TRUE) %>% 
    mutate(
    n_VE_studies_g = case_when(
      n_VE_studies > 0 & n_VE_studies < 6 ~ "01-05",
      n_VE_studies >= 6 & n_VE_studies < 11 ~ "06-10",
      n_VE_studies >= 11 & n_VE_studies < 16 ~ "11-15",
      n_VE_studies >= 16 ~ "16+",
      TRUE ~ "0"
    )
    )
  
world_ineq_gral <- st_as_sf(world_ineq_gral)

```


```{r}
#(A) COVID-19 vaccine introduction dates

map1_a<- ggplot(data = world_ineq_gral) +
  geom_sf(aes(fill = as.factor(V1_date_group)), position = "identity") + 
  scale_fill_manual(values = c("#2e73b8","#75b2d8","#9dd5ea", "#d7e0f4","#A2B1A4"),
                    labels = c("<Jan\n2021", "Jan\n2021", "Feb\n2021", ">Mar\n2021", "not\nreached"),
                    breaks = c("before jan/2021", "jan/2021", "feb/2021", "march/2021 +", "not reached"),
                    na.value = "#F0F3F0") +
  labs(
    title= "(A)",
    fill = "Month of\nintroduction") +
  theme_void()+
  guides(color = guide_legend(title.position = "bottom", nrow = 1))+
  theme(
    legend.position = "bottom",
    legend.text = element_text(size = 8),  # Adjust text size
    legend.title = element_text(size = 9),  # Adjust legend title size
    plot.title = element_text(size = 10) 
  )

map1_a

#Map 1.B) scale-up to 40% population-wide primary series coverage
map1_b<- ggplot(data = world_ineq_gral) +
  geom_sf(aes(fill = as.factor(V40_date_group)), position = "identity") + 
  scale_fill_manual(values = c("#827717","#afb42b", "#cddc39","#ecebb1","#A2B1A4"),
                    breaks = c("before Jun/2021", "Jun-Aug/2021", "Sep-Nov/2021", "Dec/2021 +", "not reached"),
                    labels = c("<Jun\n2021", "Jun-Aug\n2021", "Sep-Nov\n2021", ">Dec\n2021", "not\nreached"),
                    na.value = "#F0F3F0") +
  labs(
           title= "(B)",
    fill = "Month of\nscale-up (40%)") +
  theme_void()+ 
  theme(
    legend.position = "bottom",
    legend.text = element_text(size = 8),  # Adjust text size
    legend.title = element_text(size = 9),  # Adjust legend title size
    plot.title = element_text(size = 10) # Adjust plot title size
  ) 

map1_b

#Map 1.C: distribution of available VE studies across different countries

map1_c<- ggplot(data = world_ineq_gral) +
  geom_sf(aes(fill = n_VE_studies_g), 
          position = "identity") + 
  scale_fill_manual(values = c("#F0F0F0", "#C8E6A4", "#79C77A", "#3D8B62", "#005147"),
                    labels = c("0","01-05", "06-10", "11-15", "16+"))+ 
  labs(
    title= "(C)",
    fill = 'Number of\nVE studies') +
  theme_void() +
  theme(
    legend.position = "bottom",        # Move legend to the bottom
    legend.key.width = unit(0.5, "cm"), # Adjust the width of legend key to make it smaller
    legend.text = element_text(size = 8),  # Adjust text size
    legend.title = element_text(size = 9),  # Adjust legend title size
    plot.title = element_text(size = 10) 
  )

map1_c


map_1 <- map1_a / map1_b / map1_c
map_1


```


##########################################################################################################################
# Fig. 2. Disparities in COVID-19 vaccine introduction, scale-up and vaccine effectiveness evaluation by income quartile.#
##########################################################################################################################

```{r}

## Fig 2.a:  (A) Percentage of countries achieving vaccine introduction by 31 December 2021 (left), scale-up to 40% population-wide primary series coverage by 31 December 2021 (middle), and publication of ≥1 VE study by 07 January 2024 (right) by GNI quartile (n countries = 51, 51, 51, 50 for quartiles 4, 3, 2, 1, respectively). 

summary_milestone_achieved <- Ineq_merged %>% 
  group_by(q_GNI_per_c_WB) %>% 
  summarise(
    V1_per = round(sum(V1_dummy == "yes", na.rm = TRUE) / n() * 100, 1),
    V40_per = round(sum(V40_dummy == "yes", na.rm = TRUE) / 
                    sum(V1_dummy == "yes", na.rm = TRUE) * 100, 1),
        study_per = round(sum(study_dummy == "yes")/ sum(V1_dummy == "yes", na.rm = TRUE) * 100, 1))

summary_milestone_achieved

summary_milestone_achieved_l <- summary_milestone_achieved %>% 
    pivot_longer(
    cols = c(V1_per, V40_per, study_per),
    names_to = "milestone",
    values_to = "Value"
  )

summary_milestone_achieved_l$milestone <- factor(
  summary_milestone_achieved_l$milestone,
  levels = c("V1_per", "V40_per", "study_per")
)
summary_milestone_achieved_l

# Plot
Fig_2_a_vac_achiev <- ggplot(summary_milestone_achieved_l, aes(x = as.factor(q_GNI_per_c_WB), y = Value, fill = as.factor(q_GNI_per_c_WB))) +
  geom_bar(stat = "identity", position = "dodge") +
  scale_fill_manual(values = c("#77AADD","#44BB99","#AAAA00","#EE8866")) +
  coord_flip()+
facet_wrap(~ milestone, labeller = labeller(milestone = c(
  "V1_per" = "Vaccine introduction (31/12/2021)", 
  "V40_per" = "Vaccine scale-up (40%, 31/12/2021)",
  "study_per" = "% countries with ≥1 VE study (07/01/2024)"
))) +
  labs(title = "(A)",
       x = "",
       y = "% countries achieving milestone") +
    guides(fill = "none") +  # Remove the fill legend
  theme_minimal() +
  scale_x_discrete(labels = c("1 (lower\nincome)", "2", "3", "4 (higher\nincome)")) + 
  theme_minimal() + 
  theme(
    legend.position = "bottom",
    legend.text = element_text(size = 8),# Adjust text size
    legend.title = element_text(size = 9),  # Adjust legend title size
    plot.title = element_text(size = 10) # Adjust plot title size
  ) 

Fig_2_a_vac_achiev

# Percentage of VE studies by GNIpc GNI quartile + min pub date
summ_GNI_eff_study <- Ineq_merged %>%
  select(Country.Name,q_GNI_per_c_WB, n_VE_studies, min_VE_pub_date) %>%
  filter(!is.na(n_VE_studies)) %>%  # Exclude rows with NA in the 'n_VE_studies' variable
  group_by(
    q_GNI_per_c_WB
  ) %>%
  summarize(
    total_st_GNI = sum(n_VE_studies),
    perc_st_GNI= sum(n_VE_studies)/423*100, #fix with final value
    min_VE_pub_date= min(min_VE_pub_date)
  )
summ_GNI_eff_study

# Percentage of countries with VE studies available by GNIpc GNI quartile.
Ineq_merged_introduced <- subset(Ineq_merged,  V1_dummy=="yes")

table(Ineq_merged_introduced$q_GNI_per_c_WB)

eff_study_perc <-Ineq_merged %>%
  group_by(q_GNI_per_c_WB) %>%
  summarise(
    study = sum(study_dummy == "yes"))%>%
  mutate(
    countries_quartile=case_when(
      q_GNI_per_c_WB == 1 ~ 50,
      q_GNI_per_c_WB == 2 ~ 51,
      q_GNI_per_c_WB == 3 ~ 51,
      q_GNI_per_c_WB == 4 ~ 51
    ),
  ) %>%
  mutate(
    study_per = round(study / countries_quartile * 100, 0),
    )
eff_study_perc

```


```{r}
#Fig 2.b: Date of vaccine introduction (left; n countries = 51, 51, 51, and 50 for quartiles 4, 3, 2, 1, respectively), scale-up to 40% population-wide primary series coverage (middle; n = 49, 43, 32, 22) and first VE study publication (right; n countries= 24, 15, 7, 2) by GNI quartile. Only countries that reached the specified milestones are included. 

#Data for plot
intro_scale_delay<- Ineq_merged %>%
  mutate(
    owid_vac_F40 = as.Date(owid_vac_F40)  ) %>% 
  select(Country.Name, q_GNI_per_c_WB, min_vac_intro, owid_vac_F40, min_VE_pub_date) %>%
  filter(!is.na(min_vac_intro)) %>%
  pivot_longer(
    cols = c(min_vac_intro, owid_vac_F40, min_VE_pub_date),
    names_to = "Date",
    values_to = "Value"
  ) 

# Set factor levels
 intro_scale_delay$Date = factor(intro_scale_delay$Date, levels = unique(intro_scale_delay$Date))

#Plot
Fig_2_b_vac_delay <- ggplot(intro_scale_delay) +
  geom_boxplot(aes(x = Value, y = as.factor(q_GNI_per_c_WB)), outlier.shape = NA) +
   geom_point(
    aes(
     x = Value, 
     y = as.factor(q_GNI_per_c_WB),
      #size = studies,
      color = as.factor(q_GNI_per_c_WB)
    ),
    alpha = 0.5,
    position = position_jitterdodge(dodge.width = 0.5, jitter.width = 1) 
  )+
  scale_x_date(limits = c(as.Date("2020-12-01"), NA),
               date_labels = "%b/%Y") +
  scale_color_manual(values = c("#77AADD","#44BB99","#AAAA00","#EE8866")) +
  facet_wrap(~Date, scales = "free_x", labeller = labeller(Date = c("min_vac_intro" = "Date of vaccine introduction", "owid_vac_F40" = "Date of vaccine scale-up (40%)", "min_VE_pub_date"= "Date of first VE study")), nrow=1)+
  labs(title = "(B)",
       x = "Date", 
       y = "GNI quartile")  +
  guides(color = "none") +
  scale_y_discrete(labels = c("1 (lower\nincome)", "2", "3", "4 (higher\nincome)")) + 
  theme_minimal() + 
  theme(
    legend.position = "bottom",
    legend.text = element_text(size = 8),# Adjust text size
    legend.title = element_text(size = 9),  # Adjust legend title size
    plot.title = element_text(size = 10) # Adjust plot title size
  ) 
 
Fig_2_b_vac_delay 

# table n
table_countries_by_gni_date_wide <- intro_scale_delay %>%
  filter(!is.na(Value)) %>%
  group_by(q_GNI_per_c_WB, Date) %>%
  summarise(n_countries = n_distinct(Country.Name), .groups = "drop") %>%
  pivot_wider(names_from = Date, values_from = n_countries)

table_countries_by_gni_date_wide

summary_vacc_eff_delay <- Ineq_merged %>%
  group_by(q_GNI_per_c_WB) %>%
  summarize(
    median_value_1 = round(median(as.numeric(diff_intro_min_VE_pub), na.rm = TRUE),0),
    q25_1 = round(quantile(as.numeric(diff_intro_min_VE_pub), 0.25, na.rm = TRUE),0),
    q75_1 = round(quantile(as.numeric(diff_intro_min_VE_pub), 0.75, na.rm = TRUE),0)
  )
summary_vacc_eff_delay

#Create table with study counts and introduction dates by GNI quartile
GNI_studies <- COVID19VE_ineq_merged %>%
  group_by(q_GNI_per_c_WB) %>%
  summarise(
    min_VE_pub_date = min(as.Date(min_pub_date)),
    n_VE_studies = n_distinct(Study_id_f),
    perc_VE_studies = round(n_distinct(Study_id_f)/423*100,0)
  )
```



```{r}
# 2.C: Percentage of countries using different vaccine platforms by GNI quartile as of 07 January 2024 (n countries = 51, 51, 51, 50 for quartiles 4, 3, 2, 1, respectively). 

#Pivot longer + plat classification
owid_vaccine_country_g <- Ineq_merged %>% 
    pivot_longer(
    cols = `CIGB CIGB-66/Abdala`: `Valneva VLA2001`,
    names_to = "vaccine",
    values_to = "check",
    values_drop_na = TRUE) %>% 
  filter(
    check== 1
  ) %>% 
  select(-check) %>%
  mutate(
    vac_type = case_when(
      vaccine %in% c("BioNTech/Pfizer BNT162b2", "Moderna mRNA-1273") ~ "mRNA",
      vaccine %in% c("Oxford/AstraZeneca ChAdOx1-S", "Janssen Ad26.COV2.S", "Gamaleya Gam-COVID-Vac/Sputnik V", "Cansino Ad5-nCoV") ~ "Vectored",
      vaccine %in% c("Beijing/Sinopharm BBIBP-CorV", "Sinovac CoronaVac", "Bharat Covaxin/BBV152", "Wuhan/Sinopharm vaccine", "RIBSP Kazakhstan QazCovid-in", "Shifa Pharmed COVIran", "Erciyes TURKOVAC", "Shenzhen Kangtai KCONVAC", "Institute of Medical Biology CAMS vaccine", "Chumakov Center CoviVac", "Valneva VLA2001") ~ "Inactivated",
      vaccine %in% c("Novavax NVX-CoV2373", "CIGB CIGB-66/Abdala", "Instituto Finlay de Vacunas Soberana 02", "AZLB ZF2001", "Vector Institute EpiVacCorona", "Medigen MVC-COV1901", "Instituto Finlay de Vacunas Soberana 01", "Biological E Corbevax", "Sanofi/GSK CoV2 preS dTM", "Medicago CoVLP") ~ "Protein",
      TRUE ~ "Unknown"
    )
  ) 

unique(owid_vaccine_country_g$vaccine)

#Vaccine platform by GNI quartile
summary_vac_plat_distrib_GNI <-owid_vaccine_country_g %>% 
  select(
    Country.Code, q_GNI_per_c_WB, vac_type
  ) %>% 
  distinct() %>% 
  filter(
    !is.na(q_GNI_per_c_WB)
  ) %>%  
  group_by(q_GNI_per_c_WB, vac_type) %>% 
  summarise(total=n()) %>% 
  filter(
    vac_type=="Inactivated" |vac_type=="mRNA" |vac_type=="Vectored" |vac_type=="Protein"
  ) %>% 
  mutate(
    countries_quartile=case_when(
      q_GNI_per_c_WB == 1 ~ 50,
      q_GNI_per_c_WB == 2 ~ 51,
      q_GNI_per_c_WB == 3 ~ 51,
      q_GNI_per_c_WB == 4 ~ 51
    )
  ) %>% 
  mutate(
    vac_type_per = round(total / countries_quartile * 100, 1),
    ) 

summary_vac_plat_distrib_GNI

summary_vac_plat_distrib_GNI$vac_type <- factor(summary_vac_plat_distrib_GNI$vac_type, levels = c("mRNA", "Vectored", "Inactivated", "Protein"))

# Plot
Fig_2_c_plat <- ggplot(summary_vac_plat_distrib_GNI, aes(x = as.factor(q_GNI_per_c_WB), y = vac_type_per, fill = as.factor(q_GNI_per_c_WB))) +
  geom_bar(stat = "identity", position = "dodge") +
  coord_flip()+
  facet_wrap(~ vac_type, ncol=4) +
  labs(title = "(C)",
       x = "",
       y = " % countries using platform (07/01/2024)") +
    guides(fill = "none") +  # Remove the fill legend
  scale_fill_manual(values = c("#77AADD","#44BB99","#AAAA00","#EE8866")) +
  theme_minimal() +
  scale_x_discrete(labels = c("1 (lower\nincome)", "2", "3", "4 (higher\nincome)")) + 
  theme(
    legend.position = "bottom",
    legend.text = element_text(size = 8),  # Adjust text size
    legend.title = element_text(size = 9),  # Adjust legend title size
    plot.title = element_text(size = 10) 
  )

Fig_2_c_plat


#Global vaccine platform distribution %
summary_perc_by_plat <-owid_vaccine_country_g %>% 
  select(
    Country.Code, q_GNI_per_c_WB, vac_type
  ) %>% 
  distinct() %>% 
  filter(
    !is.na(q_GNI_per_c_WB)
  ) %>%  
  group_by(vac_type) %>% 
  summarise(total=n()) %>% 
  filter(
    vac_type=="Inactivated" |vac_type=="mRNA" |vac_type=="Vectored" |vac_type=="Protein"
  ) %>% 
  mutate(
    vac_type_per = round(total / 203 * 100, 1),
    ) 

summary_perc_by_plat
```





```{r}
# 2.D Percentage of primary series VE studies by platform as of 11 January 2024 (n studies = 423).

plat_summary <- COVID19VE_ineq_merged %>% 
  select(Study_id_f, min_pub_date, vac_type, q_GNI_per_c_WB) %>% 
  distinct() %>% 
  group_by(vac_type) %>% 
  summarise(
    min_pub_vac_type = min(as.Date(min_pub_date)),
    total_papers_vac_type= n(),
    per_papers_vac_type = n()/423*100
  )

plat_summary$vac_type <- factor(plat_summary$vac_type, levels = c("Protein subunit","Multiple", "Inactivated","Vectored","mRNA"))

#Plot
Fig_2_d_stu_plat<- ggplot(plat_summary, aes(x = vac_type, y =per_papers_vac_type , fill = vac_type)) +
  geom_bar(stat = "identity", position = "dodge") +
  coord_flip()+
  labs(title = "(D)",
       x = "",
       y = "% VE studies") +
    guides(fill = "none") +  # Remove the fill legend
  theme_minimal() +
  theme(
    legend.text = element_text(size = 8),  # Adjust text size
    legend.title = element_text(size = 9),  # Adjust legend title size
    plot.title = element_text(size = 10)
  ) +
  scale_fill_manual(values = c("#0072B2","#009E73", "#D55E00","#CC79A7", "#F0E442"))+
  scale_x_discrete(labels = c("Protein","Multiple", "Inactivated", "Vectored", "mRNA")) +
  scale_y_continuous(limits = c(0, 100), breaks = c(0, 25, 50, 75, 100))

Fig_2_d_stu_plat

#Create table with study counts and introduction dates by vaccine platform and outcome
plat_outcome_studies <- COVID19VE_ineq_merged %>%
  group_by(vac_type, outcome_final) %>%
  summarise(
    min_VE_pub_date = min(as.Date(min_pub_date)),
    n_VE_studies = n_distinct(Study_id_f),
    perc_VE_studies = round(n_distinct(Study_id_f)/423*100,0)
  )
```


```{r}
# FIg 2
Fig_2_l3 <- (Fig_2_c_plat | Fig_2_d_stu_plat) + plot_layout(widths = c(2, 1))
Fig_2 <- (Fig_2_a_vac_achiev / Fig_2_b_vac_delay / Fig_2_l3)

Fig_2

```


```{r}
# Percentage of countries with VE studies relating to different vaccine platforms by GNIpc GNI quartile. VE, vaccine effectiveness. Protein subunit with 1 effectiveness vaccine study was excluded
vac_plat_distrib <- COVID19VE_ineq_merged %>%
  select(Country.Code, q_GNI_per_c_WB, vac_type) %>%
  distinct() %>%
  group_by(q_GNI_per_c_WB, vac_type) %>%
  summarise(total=n()) %>%
  mutate(
    countries_quartile=case_when(
      q_GNI_per_c_WB == 1 ~ 50,
      q_GNI_per_c_WB == 2 ~ 51,
      q_GNI_per_c_WB == 3 ~ 51,
      q_GNI_per_c_WB == 4 ~ 51
    )
  ) %>%
  mutate(
    vac_type_per = round(total / countries_quartile * 100, 0),
    )
vac_plat_distrib

#Plot

vac_plat_distrib$vac_type <- factor(vac_plat_distrib$vac_type, levels = c("Protein subunit", "mRNA", "Vectored", "Inactivated", "Multiple"))


plot2_f<- ggplot(vac_plat_distrib[vac_plat_distrib$vac_type != "Protein subunit", ], aes(x = as.factor(q_GNI_per_c_WB), y = vac_type_per, fill = as.factor(q_GNI_per_c_WB))) +
  geom_bar(stat = "identity", position = "dodge") +
  coord_flip()+
  facet_wrap(~ vac_type, nrow = 2) +
  labs(title = "(F)",
       x = "GNI quartile",
       y = "% countries with VE studies") +
    guides(fill = "none") +  # Remove the fill legend
scale_fill_manual(values = c("#77AADD","#44BB99","#AAAA00","#EE8866")) +
  scale_x_discrete(labels = c("1 (lower\nincome)", "2", "3", "4 (higher\nincome)")) +
  theme_minimal()+
  theme(
    legend.text = element_text(size = 8),  # Adjust text size
    legend.title = element_text(size = 9),  # Adjust legend title size
    plot.title = element_text(size = 10)
  )+
  scale_y_continuous(limits = c(0, 50), breaks = c(0, 25, 50))
plot2_f

```


```{r}
#Manuscript data
# number of papers by outcome
outcome_summary <- COVID19VE_ineq_merged %>% 
  select(Study_id_f, outcome_final) %>% 
  distinct() %>% 
  group_by(outcome_final) %>% 
  summarise(
    total_papers_out= n(),
    per_papers_out = round(n()/nstudies*100,0)
  )
outcome_summary

# number of papers by variant
variant_summary <- COVID19VE_ineq_merged %>% 
  select(Study_id_f, variant_clean) %>% 
  distinct() %>% 
  group_by(variant_clean) %>% 
  summarise(
    total_papers_out= n(),
    per_papers_out = round(n()/nstudies*100,0)
  )
variant_summary
```




##############################################################################
#Supplementary Fig. 3. Distribution of vaccine effectiveness papers over time.#
###############################################################################


```{r }

# Supplementary Fig. 3. Distribution of vaccine effectiveness papers over time. Data are displayed by (A) GNI quartile (n publications =  345, 64, 9, 5 for quartiles 4, 3, 2, 1, respectively). and (B) vaccine platform (n publications = X for mRNA, X for vectored, X for inactivated, X for multiple, and X for protein).

# (A) GNI quartile 
data_F2 <- COVID19VE_ineq_merged %>% 
  select(country_clean, iso3c, min_pub_date, semester,Study_id_f, q_GNI_per_c_WB) %>% 
  distinct()

time_a<- data_F2 %>%
  ggplot(aes(x = as.Date(min_pub_date), y = as.factor(q_GNI_per_c_WB))) +
  geom_boxplot(outlier.shape = NA) +
  geom_jitter(aes(color = as.factor(q_GNI_per_c_WB), alpha = 0.5)) +
   guides(alpha = "none") +
  scale_color_manual(values = c("#77AADD","#44BB99","#AAAA00","#EE8866")) + #I dont like the scale
  labs(x = "Publication Date", y = "GNI quartile")  +
  guides(color = FALSE) +
  scale_y_discrete(labels = c("1 (lower\nincome)", "2", "3", "4 (higher\nincome)")) + 
  theme_minimal()

paper_GNI <- data_F2 %>%
  group_by(
    q_GNI_per_c_WB) %>%
  summarise(
    total_papers = n()
  ) 

paper_sem_quart <- data_F2 %>%
  group_by(
    semester, 
    q_GNI_per_c_WB) %>%
  summarise(
    total_papers = n()
  ) 
  
   
time_b <- ggplot(paper_sem_quart, aes(x = semester, y = total_papers, fill = as.factor(q_GNI_per_c_WB))) +
  geom_bar(stat = "identity") +
  labs(x = "", y = "N VE studies", fill = "GNI quartile") +
  theme_minimal() +
  theme(legend.position = "bottom", axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_fill_manual(values = c("#77AADD","#44BB99","#AAAA00","#EE8866"),
    labels = c("1 (lower\nincome)", "2", "3", "4 (higher\nincome)") ) 

```



```{r }

# (B) vaccine platform

data_F3 <- COVID19VE_ineq_merged %>% 
  select(country_clean, iso3c, min_pub_date, semester, Study_id_f, q_GNI_per_c_WB, vac_type) %>% 
  distinct() 

data_F3$vac_type <- factor(data_F3$vac_type, levels = c("Protein subunit","Multiple", "Inactivated", "Vectored", "mRNA"))


# (B) vaccine platform. In the upper panels, each point represents a study. 
timeB_a<- data_F3 %>%
  ggplot(aes(x = as.Date(min_pub_date), y = as.factor(vac_type))) +
  geom_boxplot(outlier.shape = NA) +
  geom_jitter(aes(color = as.factor(vac_type),alpha = 0.5)) +
  guides(alpha = "none") +
  scale_color_manual(values = c("#0072B2","#009E73", "#D55E00","#CC79A7", "#F0E442"))+
  labs(x = "Publication Date", y = "Vaccine platform")  +
  guides(color = FALSE) +
  scale_y_discrete(labels = c("Protein","Multiple", "Inactivated", "Vectored", "mRNA")) + 
  theme_minimal()

paper_sem_plat <- data_F3 %>%
  group_by(semester, vac_type) %>%
  summarise(
    total_papers = n()
  ) 

paper_plat <- data_F3 %>%
  group_by(vac_type) %>%
  summarise(
    total_papers = n()
  ) 
  
timeB_b <- ggplot(paper_sem_plat, aes(x = semester, y = total_papers, fill = as.factor(vac_type))) +
  geom_bar(stat = "identity") +
  labs(x = "", y = "N VE studies", fill = "Vaccine platform") +
  theme_minimal() +
  theme(legend.position = "bottom", axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_fill_manual(values = c("#0072B2","#009E73", "#D55E00","#CC79A7", "#F0E442"),
    labels = c("Protein","Multiple", "Inactivated", "Vectored", "mRNA"))


# Extended Data Fig. 3. Distribution of vaccine effectiveness papers over time.
combined_plotB <- time_a/time_b/timeB_a/timeB_b
combined_plotB 



vac_type_sum2 <- data_F3 %>%
  group_by(vac_type) %>%
  summarise(min_date = min(as.Date(min_pub_date)),
            ) 



median_labels <- data_F3 %>%
  group_by(vac_type) %>%
  summarize(
    median_value = as.Date(median(as.numeric(min_pub_date), na.rm = TRUE), origin = "1970-01-01"),
    q25 = as.Date(quantile(as.numeric(min_pub_date), 0.25, na.rm = TRUE), origin = "1970-01-01"),
    q75 = as.Date(quantile(as.numeric(min_pub_date), 0.75, na.rm = TRUE), origin = "1970-01-01")
  )
```

##########################################################################################################
# Table 1. Characteristics of primary series vaccine effectiveness studies reporting on country-specific# estimates.
########################################################################################################


```{r}
#Table 1. Characteristics of primary series vaccine effectiveness studies reporting on country-specific estimates.

# vaccine study distribution
table_vacc <- COVID19VE_ineq_merged %>%
  distinct(Study_id_f, clean_vaccine) %>% 
  mutate(
    total_study=n_distinct(Study_id_f) 
  ) %>% 
  group_by(clean_vaccine, total_study) %>%
  mutate(n = n(),
         `n(%)` = paste0( n, " (", round(n / total_study * 100, 0),"%)" )) %>% 
  ungroup() %>% 
  select(clean_vaccine, `n(%)`) %>% 
  distinct()

table_vac2 <- COVID19VE_ineq_merged %>%
  distinct(Study_id_f, vac_type, clean_vaccine) %>% 
  group_by(vac_type) %>% 
  mutate(
    total_study=n_distinct(Study_id_f)) %>% 
  ungroup() %>% 
  group_by(vac_type, clean_vaccine,total_study) %>% 
 mutate(n = n(),
         `n(%)` = paste0( n, " (", round(n / total_study * 100, 0),"%)" )) %>% 
  ungroup() %>% 
  select(vac_type, clean_vaccine, `n(%)`) %>% 
  distinct() %>% 
  pivot_wider(names_from = vac_type, values_from = `n(%)`)
  
combined_Vac <- left_join(table_vacc, table_vac2)


# vaccine platform study distribution
table_plat_vac <- COVID19VE_ineq_merged %>%
  distinct(Study_id_f, vac_type) %>% 
  mutate(
    total_study=n_distinct(Study_id_f) 
  ) %>% 
  group_by(vac_type, total_study) %>%
  mutate(n = n(),
         `n(%)` = paste0( n, " (", round(n / total_study * 100, 0),"%)" )) %>% 
  ungroup() %>% 
  select(vac_type, `n(%)`) %>% 
  distinct()


#variant of interest
table_VoI1 <- COVID19VE_ineq_merged %>%
  distinct(Study_id_f, variant_clean) %>% 
  mutate(
    total_study=n_distinct(Study_id_f) 
  ) %>% 
  group_by(variant_clean, total_study) %>% 
   mutate(n = n(),
         `n(%)` = paste0( n, " (", round(n / total_study * 100, 0),"%)" )) %>% 
  ungroup() %>% 
  select(variant_clean, `n(%)`) %>% 
  distinct()

table_VoI2 <- COVID19VE_ineq_merged %>%
  distinct(Study_id_f, vac_type, variant_clean) %>% 
  group_by(vac_type) %>% 
  mutate(
    total_study=n_distinct(Study_id_f)) %>% 
  ungroup() %>% 
  group_by(vac_type, variant_clean,total_study) %>% 
 mutate(n = n(),
         `n(%)` = paste0( n, " (", round(n / total_study * 100, 0),"%)" )) %>% 
  ungroup() %>% 
  select(vac_type, variant_clean, `n(%)`) %>% 
  distinct() %>% 
  pivot_wider(names_from = vac_type, values_from = `n(%)`)
  
combined_VoI <- left_join(table_VoI1, table_VoI2)

#outcome
table_outcome1 <- COVID19VE_ineq_merged %>%
  distinct(Study_id_f, outcome_final) %>% 
  mutate(
    total_study=n_distinct(Study_id_f) 
  ) %>% 
  group_by(outcome_final, total_study) %>% 
   mutate(n = n(),
         `n(%)` = paste0( n, " (", round(n / total_study * 100, 0),"%)" )) %>% 
  ungroup() %>% 
  select(outcome_final, `n(%)`) %>% 
  distinct()
table_outcome2 <- COVID19VE_ineq_merged %>%
  distinct(Study_id_f, vac_type, outcome_final) %>% 
  group_by(vac_type) %>% 
  mutate(
    total_study=n_distinct(Study_id_f)) %>% 
  ungroup() %>% 
  group_by(vac_type, outcome_final,total_study) %>% 
 mutate(n = n(),
         `n(%)` = paste0( n, " (", round(n / total_study * 100, 0),"%)" )) %>% 
  ungroup() %>% 
  select(vac_type, outcome_final, `n(%)`) %>% 
  distinct() %>% 
  pivot_wider(names_from = vac_type, values_from = `n(%)`)
  
combined_outcome <- left_join(table_outcome1, table_outcome2)

#GNI
table_GNI1 <- COVID19VE_ineq_merged %>%
  distinct(Study_id_f, q_GNI_per_c_WB) %>% 
  mutate(
    total_study=n_distinct(Study_id_f) 
  ) %>% 
  group_by(q_GNI_per_c_WB, total_study) %>% 
   mutate(n = n(),
         `n(%)` = paste0( n, " (", round(n / total_study * 100, 0),"%)" )) %>% 
  ungroup() %>% 
  select(q_GNI_per_c_WB, `n(%)`) %>% 
  distinct()
table_GNI2 <- COVID19VE_ineq_merged %>%
  distinct(Study_id_f, vac_type, q_GNI_per_c_WB) %>% 
  group_by(vac_type) %>% 
  mutate(
    total_study=n_distinct(Study_id_f)) %>% 
  ungroup() %>% 
  group_by(vac_type, q_GNI_per_c_WB,total_study) %>% 
 mutate(n = n(),
         `n(%)` = paste0( n, " (", round(n / total_study * 100, 0),"%)" )) %>% 
  ungroup() %>% 
  select(vac_type, q_GNI_per_c_WB, `n(%)`) %>% 
  distinct() %>% 
  pivot_wider(names_from = vac_type, values_from = `n(%)`)
  
combined_GNI <- left_join(table_GNI1, table_GNI2)
```

<!-- #### Clean supplementary dataset -->
<!-- ```{r} -->
<!-- supp_data <- Ineq_merged %>% -->
<!--   select( -->
<!--     Country.Name:source_website, -->
<!--     mRNA:Protein, -->
<!--     min_VE_pub_date:diff_intro_min_VE_pub -->
<!--   ) %>% -->
<!--   select( -->
<!--     -c(Economy, `Income group`, perc_VE_studies, study_dummy,  -->
<!--        Region, Country.Code, owid_who, diff_V40_VEpub), -->
<!--     -starts_with("V1"), -starts_with("V4"), -starts_with("V7") -->
<!--   ) %>% -->
<!--   relocate(iso3c, .after = Country.Name) %>% -->
<!--   relocate(continent, .after = iso3c) %>% -->
<!--   relocate(q_GNI_per_c_WB, .after = GNI_per_c_WB) %>% -->
<!--   rename(country = Country.Name, -->
<!--          iso_code = iso3c, -->
<!--          worldbank_GNI_per_capita = GNI_per_c_WB, -->
<!--          worldbank_GNI_quartile = q_GNI_per_c_WB, -->
<!--          worldbank_GNI_year = year_GNI, -->
<!--          owid_vaccines = vaccines, -->
<!--          owid_last_observation_date = last_observation_date, -->
<!--          any_mRNA = mRNA, -->
<!--          any_vectored = Vectored, -->
<!--          any_inactivated = Inactivated, -->
<!--          any_protein = Protein, -->
<!--          owid_source_name = source_name, -->
<!--          owid_source_website = source_website, -->
<!--          ivac_min_VE_pub_date = min_VE_pub_date, -->
<!--          ivac_n_VE_studies = n_VE_studies, -->
<!--          diff_intro_min_VE_pub = diff_intro_min_VE_pub) %>% -->
<!--   relocate(diff_intro_min_VE_pub, .after = ivac_n_VE_studies) -->




<!-- write.csv(supp_data, "Data/Processed/supplementary_data.csv") -->
<!-- ``` -->
#############################################################################################################
# Fig. 3. Correlation between COVID-19 vaccine implementation metrics and per capita Gross National Income.# 
###########################################################################################################
```{r}
# Fig. 3. Correlation between COVID-19 vaccine implementation metrics and per capita Gross National Income. 
# Build table
scatter_milestone_country <- Ineq_merged %>%
  select(Country.Name, GNI_per_c_WB, q_GNI_per_c_WB, min_vac_intro, owid_vac_F40, min_VE_pub_date) %>%
  mutate(
    owid_vac_F40 = as.Date(owid_vac_F40)
  ) %>% 
  pivot_longer(
    cols = c(min_vac_intro, owid_vac_F40, min_VE_pub_date),
    names_to = "milestone",
    values_to = "date"
  )

scatter_milestone_country$milestone = factor(scatter_milestone_country$milestone, levels=c("min_vac_intro" , "owid_vac_F40" , "min_VE_pub_date"))

#Correlation test
spearman_corr0 <- scatter_milestone_country %>%
  filter(!is.na(date)) %>% 
  group_by(milestone) %>%
  summarise(
    cor_val = round(cor.test(GNI_per_c_WB, as.numeric(date), method = "spearman")$estimate, 2),
    p_value = cor.test(GNI_per_c_WB, as.numeric(date), method = "spearman")$p.value,
    nrow = sum(!is.na(as.numeric(date)) & !is.na(GNI_per_c_WB))
  )

#Label
label_positions <- scatter_milestone_country %>%
  group_by(milestone) %>%
  summarise(
    x_pos = max(log(GNI_per_c_WB), na.rm = TRUE),
    y_pos = max(date, na.rm = TRUE),
    .groups = "drop"
  )

#Add correlation (+labels) to the long data
spearman_corr <- spearman_corr0 %>%
  left_join(label_positions, by = "milestone") %>%
  mutate(label = paste0("corr = ", cor_val))



#Plot
ED_F0_scatterplot_GNI_milestone <- ggplot(scatter_milestone_country) +
  geom_point(aes(x = log(GNI_per_c_WB), y = date, color = as.factor(q_GNI_per_c_WB)), alpha = 0.8) +
  geom_smooth(aes(x = log(GNI_per_c_WB), y = date),
              color = "black", linetype = "dotted", size = 0.5,     
              se = TRUE, alpha = 0.3) +
  geom_text(data = spearman_corr,
            aes(x = x_pos, y = y_pos, label = label),
            inherit.aes = FALSE, hjust = 1, vjust = 1, size = 3) +
  facet_wrap(~ milestone, 
             scales = "free", 
             labeller = labeller(
               milestone = c("min_vac_intro" = "Vaccine introduction",
                             "owid_vac_F40" = "Vaccine scale-up (40%)",
                             "min_VE_pub_date"= "1st VE study publication"))) + 
  scale_y_date(limits = c(as.Date("2020-12-01"), NA),
               date_labels = "%b/%Y") +
  labs(
    x = "GNI per capita (log)",
    y = "Date",
    color = "GNI quartile"
  ) +
  theme_minimal() +
  theme(
    strip.text = element_text(size = 10),
    legend.position = "bottom"
  ) +
  scale_color_manual(values = c("#77AADD","#44BB99","#AAAA00","#EE8866"),
                     labels = c("1 (lower\nincome)", "2", "3", "4 (higher\nincome)"))

ED_F0_scatterplot_GNI_milestone


```




```{r}
# Supplementary table Intro+Scale-up
ineq_intro_scale_VE <- Ineq_merged %>%
  select(
    Country.Name,
    Country.Code,
    Region,
    continent,
    GNI_per_c_WB,
    year_GNI,
    q_GNI_per_c_WB,
    min_vac_intro,
    owid_who,
    owid_vac_F40,
    owid_vac_F70,
    diff_intro_F40,
    diff_intro_F70,
    vaccines,
    last_observation_date,
    mRNA:Protein,
    source_name,
    source_website,
    min_VE_pub_date,
    diff_intro_min_VE_pub,
    n_VE_studies
  ) %>%
  rename(
  country = Country.Name,
  iso_code = Country.Code,
  continent = continent,
  worldbank_GNI_per_capita = GNI_per_c_WB,
  worldbank_GNI_year = year_GNI,
  worldbank_GNI_quartile = q_GNI_per_c_WB,
  min_vac_intro = min_vac_intro,
  min_vac_intro_source = owid_who,
  owid_vac_F40 = owid_vac_F40,
  owid_vac_F70 = owid_vac_F70,
  diff_intro_F40 = diff_intro_F40,
  diff_intro_F70 = diff_intro_F70,
  owid_vaccines = vaccines,
  owid_last_observation_date = last_observation_date,
  any_mRNA = mRNA,
  any_vectored = Vectored,
  any_inactivated = Inactivated,
  any_protein = Protein,
  owid_source_name = source_name,
  owid_source_website = source_website,
  ivac_min_VE_pub_date = min_VE_pub_date,
  diff_intro_min_VE_pub = diff_intro_min_VE_pub,
  ivac_n_VE_studies = n_VE_studies
)

curated_dataset_ineq_intro_scale_VE_prior_group<- left_join(ineq_intro_scale_VE, Ineq_merged_prior_group, by = c("iso_code"= "Country.Code")) %>% 
  mutate(
    who_unicef_perc_old = who_unicef_perc_old*100,
    who_unicef_perc_HCW = who_unicef_perc_HCW*100) %>% 
  rename(who_unicef_perc_hcw = who_unicef_perc_HCW)

write.csv(curated_dataset_ineq_intro_scale_VE_prior_group, "Data/Curated_datasets/Ineq_merged_with_vaccine_intro_scale_plat_priority_groups_VE.csv")

```




```{r, session info}

print(sessionInfo())

```
