==> Scripts/8 Meta-regression summary.Rmd <==
---
title: "8 Meta-regression summary"
author: "Martina Pesce & Edward P. K. Parker"
date: "2023-10-12"
output: html_document
editor_options: 
  chunk_output_type: console
---

#### Inputs
- Processed/COVID19VE_ineq_merged_filtered_eligible_selected_metareg_var.xlsx: Processed meta-regression data

#### Outputs
- Meta-regression model results and plots for selected vaccine
- Fig. 5. Meta-regression of primary series vaccine effectiveness against severe COVID-19 by variant context and income status.


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```



```{r}
library(tidyverse)
library(metafor)
library(forestplot)
library(patchwork)
library(scales)

VEmetareg0 <- read_csv("Data/Processed/COVID19VE_ineq_merged_filtered_eligible_selected_metareg_var.csv")
VEmetareg <- VEmetareg0 %>% 
  filter(include=="yes")

# summarise study counts
VEmetareg %>% group_by(variant_clean, clean_vaccine, q_GNI_per_c_WB) %>%
  summarise(n = length(unique(country_clean)))
# 1 country:
# BNT162b2 delta GNI 2
# mRNA-1273 delta GNI 2
# BNT162b2 omicron GNI 2
# mRNA-1273 omicron GNI 2

# 2 country:
# ChAdOx delta GNI 1/2
# ChAdOx omicron GNI 2/3
# mRNA-1273 omicron GNI 3
# mRNA-1273 delta GNI 3
# Recode levels
VEmetareg$clean_vaccine[VEmetareg$clean_vaccine=="BioNTech/Pfizer BNT162b2"] = "BNT162b2"
VEmetareg$clean_vaccine[VEmetareg$clean_vaccine=="Oxford/AstraZeneca ChAdOx1-S"] = "ChAdOx1-S"
VEmetareg$clean_vaccine[VEmetareg$clean_vaccine=="Moderna mRNA-1273"] = "mRNA-1273"
VEmetareg$clean_vaccine = factor(VEmetareg$clean_vaccine, levels = c("BNT162b2", "ChAdOx1-S", "mRNA-1273"))

metareg_collated <- rbind(
  read_csv("Data/Processed/meta_summary_BNT162b2.csv"),
  read_csv("Data/Processed/meta_summary_ChAdOx1-S.csv"),
  read_csv("Data/Processed/meta_summary_mRNA-1273.csv")
) 
metareg_collated$clean_vaccine = factor(metareg_collated$clean_vaccine, levels = c("BNT162b2", "ChAdOx1-S", "mRNA-1273"))

metareg_collated$`N countries` = ">2"

metareg_collated$`N countries`[metareg_collated$clean_vaccine == "BNT162b2" & 
                          metareg_collated$GNI %in% c(2)] = "1-2"

metareg_collated$`N countries`[metareg_collated$clean_vaccine == "mRNA-1273" & 
                          metareg_collated$GNI %in% c(2, 3)] = "1-2"

metareg_collated$`N countries`[metareg_collated$VoI == "delta" & 
                          metareg_collated$clean_vaccine == "ChAdOx1-S" & 
                          metareg_collated$GNI %in% c(1,2)] = "1-2"

metareg_collated$`N countries`[metareg_collated$VoI == "omicron" & 
                          metareg_collated$clean_vaccine == "ChAdOx1-S" & 
                          metareg_collated$GNI %in% c(2,3)] = "1-2"

# metareg_collated$`N countries`[metareg_collated$VoI == "omicron" & 
#                           metareg_collated$clean_vaccine == "mRNA-1273" & 
#                           metareg_collated$GNI %in% c(3)] = "1-2"


## set variant levels
metareg_collated$VoI = factor(metareg_collated$VoI, levels=c("delta", "omicron"))
metareg_collated$variant_clean = metareg_collated$VoI
VEmetareg$variant_clean = factor(VEmetareg$variant_clean, levels=c("delta", "omicron"))

## set GNI perc c levels
metareg_collated$GNI = factor(metareg_collated$GNI, levels=c("1", "2", "3", "4"))
VEmetareg$q_GNI_per_c_WB = factor(VEmetareg$q_GNI_per_c_WB, levels=c("1", "2", "3", "4"))

## reverse log10 transformation
reverselog_trans <- function(base = exp(1)) {
    trans <- function(x) -log(x, base)
    inv <- function(x) base^(-x)
    trans_new(paste0("reverselog-", format(base)), trans, inv, 
              log_breaks(base = base), 
              domain = c(1e-100, Inf))
}

# --- labels n=xx por faceta y GNI ---
labels_df <- VEmetareg %>%
  filter(!is.na(RR)) %>%
  transmute(
    variant_clean = fct_relevel(variant_clean, c("delta","omicron")),
    clean_vaccine,
    GNI = as.integer(q_GNI_per_c_WB)
  ) %>%
  count(variant_clean, clean_vaccine, GNI, name = "n_points") %>%
  complete(variant_clean, clean_vaccine, GNI = 1:4, fill = list(n_points = 0L)) %>%
  mutate(label = paste0("n=", n_points))


# Fig. 5
fig_5 <- ggplot() +
  geom_point(
    data = metareg_collated,
    aes(x = GNI, y = pred, color = as.factor(GNI), shape = `N countries`),
    size = 2.5
  ) +
  geom_errorbar(
    data = metareg_collated,
    aes(x = GNI, y = pred, color = as.factor(GNI), ymin = ci.lb, ymax = ci.ub),
    width = 0
  ) +
  geom_point(
    data = VEmetareg,
    aes(x = q_GNI_per_c_WB, y = RR, size = 1 / SE.ln.RR ^ 2, color = as.factor(q_GNI_per_c_WB)),
    alpha = 0.25
  ) +
  facet_grid(variant_clean ~ clean_vaccine, drop = FALSE) +
  coord_flip(clip = "off") +
  scale_x_discrete(labels = c("1 (lower\nincome)", "2" ,"3","4 (higher\nincome)")) +
  scale_shape_manual(values = c(15, 0)) +
  scale_color_manual(
    name = "GNI quartile",
    values = c("#77AADD", "#44BB99", "#AAAA00", "#EE8866"),
    labels = c("1 (lower\nincome)", "2", "3", "4 (higher\nincome)")
  ) +
  scale_y_continuous(
    breaks = c(0.01, 0.1, 0.5, 1),
    labels = c(99, 90, 50, 0),
    trans = reverselog_trans(10),
    expand = expansion(mult = c(0, 0.15))
  ) +
  guides(size = "none", colour = "none", shape = guide_legend(reverse = TRUE)) +
  geom_hline(yintercept = 1, linetype = "dashed", color = "black") +
  labs(x = "GNI quartile", y = "Vaccine effectiveness (%)") +
  theme_minimal() +
  theme(
    legend.position = "right",
    axis.text.x = element_text(angle = 45, hjust = 1),
    axis.text.y = element_text(size = 9),
    strip.text = element_text(size = 10),
    panel.spacing = unit(0.7, "lines")
  ) +
  geom_text(
    data = labels_df,
    aes(x = GNI, label = label),
    y = Inf,
    hjust = 1, vjust = 0,
    size = 3,
    inherit.aes = FALSE,
    color = "#666666"
  )

fig_5
```
